Analysis.Parsimony.Additive.Internal

Description

Standard Additive character analysis (cost and medians).

This only works on static characters, and due to the traversal, only one character will be received at a time.

Assumes binary trees.

Note that this is the same procedure as for continuous characters, but with costs of Word, not Double.

Documentation

additivePostOrder :: (DiscreteCharacterMetadata d, RangedCharacterDecoration d c, DiscreteCharacterMetadata d', RangedExtensionPostorder d' c) => d -> [d'] -> d'

Used on the post-order (i.e. first) traversal. Applies appropriate logic to internal node and leaf node cases.

initializeLeaf :: (DiscreteCharacterMetadata d, RangedCharacterDecoration d c, RangedExtensionPostorder d' c) => d -> d'

Initializes a leaf node by copying its current value into its preliminary state. Gives it a minimum cost of 0.

Used on the postorder pass.

updatePostOrder :: (RangedCharacterDecoration d c, DiscreteCharacterMetadata d', RangedExtensionPostorder d' c) => d -> NonEmpty d' -> d'

Updates the character on the parent of two child nodes to become an AdditiveOptimizationDecoration. Determines the cost by adding the cost of the intersection of the two child nodes, then summing that value with the costs of the two children. The preliminary value of the character is the intersection of the two child intervals.

Used on the postorder pass.

additivePreOrder :: (DiscreteCharacterMetadata d, RangedExtensionPostorder d c, RangedExtensionPreorder d' c) => d -> [(Word, d')] -> d'

Used on the pre-order (i.e. second) traversal. Applies appropriate logic to root node, internal node, and leaf node cases.

finalizeLeaf :: (RangedExtensionPreorder d' c, DiscreteCharacterMetadata d, RangedPostorderDecoration d c) => d -> d'

Finalize a leaf node on a pre-order traversal. Set the preliminary interval as the final interval of the leaf decoration.

determineFinalState :: (Ord (Bound c), RangedExtensionPostorder d c, RangedDecorationOptimization d' c) => d -> d' -> Range (Bound c)

Uses the preliminary intervals of a node, its parents, and its children. Follows the three rules of Fitch, modified for additive characters:

If the intersection of the current node's character with its parent == the parent interval, use the parent interval

If the union of those two characters == the child, then use the child

Otherwise, find the intersections of the parent and each of the children, union them, then union that with the parent.

Used on the preorder pass.