Analysis.Parsimony.Dynamic.DirectOptimization

Description

Sankoff character analysis (cost and median)

This only works on static characters, and due to the traversal, only one character will be received at a time.

Assumes binary trees.

Documentation

data DenseTransitionCostMatrix

Exposed wrapper for C allocated cost matrix structs.

directOptimizationPreOrder :: (DirectOptimizationPostOrderDecoration d c, EncodedAmbiguityGroupContainer c, Exportable (Element c), Show (Element c)) => PairwiseAlignment c -> d -> [(Word, DynamicDecorationDirectOptimization c)] -> DynamicDecorationDirectOptimization c

The pre-order scoring logic for dynamic characters.

Parameterized over a PairwiseAlignment function to allow for different atomic alignments depending on the character's metadata.

directOptimizationPostOrder :: SimpleDynamicDecoration d c => PairwiseAlignment c -> d -> [DynamicDecorationDirectOptimizationPostOrderResult c] -> DynamicDecorationDirectOptimizationPostOrderResult c

The post-order scoring logic for dynamic characters.

Parameterized over a PairwiseAlignment function to allow for different atomic alignments depending on the character's metadata.

foreignPairwiseDO

Arguments

:: (EncodableDynamicCharacter s, Exportable s)

=> s
First dynamic character

-> s
Second dynamic character

-> DenseTransitionCostMatrix
Structure defining the transition costs between character states

-> (Word, s, s, s, s)
The ungapped character derived from the the input characters' N-W-esque matrix traceback

Align two dynamic characters using an FFI call for more efficient computation on small alphabet sizes.

Requires a pre-generated DenseTransitionCostMatrix from a call to generateDenseTransitionCostMatrix defining the alphabet and transition costs.

foreignThreeWayDO

Arguments

:: (EncodableDynamicCharacter s, Exportable s)

=> s
First dynamic character

-> s
Second dynamic character

-> s
Third dynamic character

-> Int
Mismatch cost

-> Int
Gap open cost

-> Int
Indel cost

-> DenseTransitionCostMatrix
Structure defining the transition costs between character states

-> (Word, s, s, s, s, s)
The ungapped character derived from the the input characters' N-W-esque matrix traceback

Align three dynamic characters using an FFI call for more efficient computation on small (or smallish) alphabet sizes.

Requires a pre-generated DenseTransitionCostMatrix from a call to generateDenseTransitionCostMatrix defining the alphabet and transition costs.

generateDenseTransitionCostMatrix

Arguments

:: Word
The gap open cost. A zero value indicates non-affine alignment context

-> Word
The character alphabet size

-> (Word -> Word -> Word)
The function defining the cost to transition between two symbols

-> DenseTransitionCostMatrix

Generate the 2D and 3D dense TCM matricies used for FFI calls to foreignPairwiseDO and foreignThreeWayDO.

naiveDO

Arguments

:: DOCharConstraint s

=> s
First dynamic character

-> s
Second dynamic character

-> (Word -> Word -> Word)
Structure defining the transition costs between character states

-> (Word, s, s, s, s)
The cost of the alignment

The ungapped character derived from the the input characters' N-W-esque matrix traceback

The gapped character derived from the the input characters' N-W-esque matrix traceback

The gapped alignment of the first input character when aligned with the second character

The gapped alignment of the second input character when aligned with the first character

Performs a naive direct optimization. Takes in two characters to run DO on and a metadata object Returns an assignment character, the cost of that assignment, the assignment character with gaps included, the aligned version of the first input character, and the aligned version of the second input character. The process for this algorithm is to generate a traversal matrix, then perform a traceback.

naiveDOConst :: DOCharConstraint s => s -> s -> (Word -> Word -> Word) -> (Word, s, s, s, s)

The same as naiveDO except that the "cost structure" parameter is ignored. Instead a constant cost is used.

naiveDOMemo :: DOCharConstraint s => s -> s -> OverlapFunction (Element s) -> (Word, s, s, s, s)

The same as naiveDO except that the "cost structure" parameter is assumed to be a memoized overlap function.

selectDynamicMetric :: (EncodableDynamicCharacter c, Exportable c, Exportable (Element c), HasDenseTransitionCostMatrix dec (Maybe DenseTransitionCostMatrix), HasSparseTransitionCostMatrix dec MemoizedCostMatrix, Ord (Element c)) => dec -> c -> c -> (Word, c, c, c, c)

Select the most appropriate direct optimization metric implementation.

ukkonenDO :: DOCharConstraint s => s -> s -> OverlapFunction (Element s) -> (Word, s, s, s, s)

O( (n - m + 1 ) * log(n - m + 1) ), n >= m

Compute the alignment of two dynamic characters and the median states by using Ukkone's string edit distance algorthim to improve space and time complexity.