Analysis.Parsimony.Dynamic.DirectOptimization.Pairwise

Description

Pairwise direct optimization alignment functions using a variety of techniques.

Documentation

data DenseTransitionCostMatrix

Exposed wrapper for C allocated cost matrix structs.

type OverlapFunction e = e -> e -> (e, Word)

A generalized function represention the "overlap" between dynamic character elements, supplying the corresponding median and cost to align the two characters.

filterGaps :: EncodableDynamicCharacter s => s -> s

Strips the gap elements from the supplied character.

If the character contains only gaps, a missing character is returned.

foreignPairwiseDO

Arguments

:: (EncodableDynamicCharacter s, Exportable s)

=> s
First dynamic character

-> s
Second dynamic character

-> DenseTransitionCostMatrix
Structure defining the transition costs between character states

-> (Word, s, s, s, s)
The ungapped character derived from the the input characters' N-W-esque matrix traceback

Align two dynamic characters using an FFI call for more efficient computation on small alphabet sizes.

Requires a pre-generated DenseTransitionCostMatrix from a call to generateDenseTransitionCostMatrix defining the alphabet and transition costs.

foreignThreeWayDO

Arguments

:: (EncodableDynamicCharacter s, Exportable s)

=> s
First dynamic character

-> s
Second dynamic character

-> s
Third dynamic character

-> Int
Mismatch cost

-> Int
Gap open cost

-> Int
Indel cost

-> DenseTransitionCostMatrix
Structure defining the transition costs between character states

-> (Word, s, s, s, s, s)
The ungapped character derived from the the input characters' N-W-esque matrix traceback

Align three dynamic characters using an FFI call for more efficient computation on small (or smallish) alphabet sizes.

Requires a pre-generated DenseTransitionCostMatrix from a call to generateDenseTransitionCostMatrix defining the alphabet and transition costs.

generateDenseTransitionCostMatrix

Arguments

:: Word
The gap open cost. A zero value indicates non-affine alignment context

-> Word
The character alphabet size

-> (Word -> Word -> Word)
The function defining the cost to transition between two symbols

-> DenseTransitionCostMatrix

Generate the 2D and 3D dense TCM matricies used for FFI calls to foreignPairwiseDO and foreignThreeWayDO.

getOverlap :: EncodableStreamElement e => e -> e -> (Word -> Word -> Word) -> (e, Word)

Memoized wrapper of the overlap function

minimalChoice :: (Bits b, Foldable1 t, Ord c) => t (b, c) -> (b, c)

Given a structure of unambiguous character elements and costs, calculates the least costly intersection of unambiguous character elements and the cost of that intersection.

naiveDO

Arguments

:: DOCharConstraint s

=> s
First dynamic character

-> s
Second dynamic character

-> (Word -> Word -> Word)
Structure defining the transition costs between character states

-> (Word, s, s, s, s)
The cost of the alignment

The ungapped character derived from the the input characters' N-W-esque matrix traceback

The gapped character derived from the the input characters' N-W-esque matrix traceback

The gapped alignment of the first input character when aligned with the second character

The gapped alignment of the second input character when aligned with the first character

Performs a naive direct optimization. Takes in two characters to run DO on and a metadata object Returns an assignment character, the cost of that assignment, the assignment character with gaps included, the aligned version of the first input character, and the aligned version of the second input character. The process for this algorithm is to generate a traversal matrix, then perform a traceback.

naiveDOConst :: DOCharConstraint s => s -> s -> (Word -> Word -> Word) -> (Word, s, s, s, s)

The same as naiveDO except that the "cost structure" parameter is ignored. Instead a constant cost is used.

naiveDOMemo :: DOCharConstraint s => s -> s -> OverlapFunction (Element s) -> (Word, s, s, s, s)

The same as naiveDO except that the "cost structure" parameter is assumed to be a memoized overlap function.

ukkonenDO :: DOCharConstraint s => s -> s -> OverlapFunction (Element s) -> (Word, s, s, s, s)

O( (n - m + 1 ) * log(n - m + 1) ), n >= m

Compute the alignment of two dynamic characters and the median states by using Ukkone's string edit distance algorthim to improve space and time complexity.