Analysis.Parsimony.Fitch.Internal

Description

Standard Fitch (non-additive) character analysis (cost and medians).

This only works on static characters, and due to the traversal, only one character will be received at a time.

Assumes binary trees.

Documentation

fitchPostOrder :: DiscreteCharacterDecoration d c => d -> [FitchOptimizationDecoration c] -> FitchOptimizationDecoration c

Used on the post-order (i.e. first) traversal.

fitchPreOrder :: EncodableStaticCharacter c => FitchOptimizationDecoration c -> [(Word, FitchOptimizationDecoration c)] -> FitchOptimizationDecoration c

Used on the pre-order (i.e. second) traversal.

updatePostOrder :: DiscreteCharacterDecoration d c => d -> NonEmpty (FitchOptimizationDecoration c) -> FitchOptimizationDecoration c

Used in first, postorder, pass. Take in parent and two child nodes. Using the child preliminary decorations, calculate the preliminary character state for the parent node. In addition, calculate the cost of assigning that character state to the parent.

initializeLeaf :: DiscreteCharacterDecoration d c => d -> FitchOptimizationDecoration c

A leaf has cost 0 and its preliminary character state is also its final character state. Its "child preliminary medians" are empty lists.

determineFinalState :: EncodableStaticCharacter c => FitchOptimizationDecoration c -> FitchOptimizationDecoration c -> FitchOptimizationDecoration c

Using the preliminary state of the current node, as well as the preliminary states of its parent and sibling, compute the final state of the character using Fitch's ordered rules.