Analysis.Parsimony.Sankoff.Internal

Description

Sankoff character analysis (cost and median)

This only works on static characters, and due to the traversal only one character will be received at a time.

Goloboff's algorithm relies on computing the "extra cost" for each non- optimal state assignment for each node. There's a preliminary extra cost, which is the difference between the assignment cost for this state on this node, and a final extra cost, which is the total extra cost when recursing over the whole tree that this state assignment implies. This involves, then, computing extra costs even for states that are non- optimal

Assumes binary trees.

Documentation

sankoffPostOrder :: DiscreteCharacterDecoration d c => d -> [SankoffOptimizationDecoration c] -> SankoffOptimizationDecoration c

Used on the post-order (i.e. first) traversal.

sankoffPreOrder :: EncodableStaticCharacter c => SankoffOptimizationDecoration c -> [(Word, SankoffOptimizationDecoration c)] -> SankoffOptimizationDecoration c

Used on the pre-order (i.e. second) traversal.

Either calls initializeDirVector on root or updateDirectionalMins. Needs to determine which child it's updating, then sends the appropriate minlist to updateDirectionalMins.

initializeCostVector :: DiscreteCharacterDecoration d c => d -> SankoffOptimizationDecoration c

Before post-order can actually occur, must initialize leaf vectors with values as such:

Given \(n\) character states, for a given character \(i_c\) on leaf \(i\), there are \(2^n - 1\) possible characters, including ambiguous characters. For extant character states \(i_{c_x}\) on the leaf, and for each possible character state, if that character state is extant on the leaf, give it an initial cost of 0, otherwise, ∞.

updateCostVector :: DiscreteCharacterDecoration d c => d -> NonEmpty (SankoffOptimizationDecoration c) -> SankoffOptimizationDecoration c

Used on the post-order (i.e. first) traversal, moving from leaves to root.

Given current node and its children, does actual calculation of new node values for each character state:

\(\forall a, minCosts_a = \forall b, \(min (transitionCost(a, b) + leftChild_{minCost(b)} + rightChild_{minCost(b)}\). (This is a tuple of lists), where

\(\forall a, minCost(a) = min(minCosts_a)\).

\(nodeMin = min(minCost)\)

In order to run Goloboff’s Sankoff traversal optimizations, computes:

\(\forall a, preliminaryExtraCost(a) = \(minCost(a) - nodeMin\)

\(\beta(a, b) = min_{a,b} (transitionCost (a, b) + preliminaryExtraCost(b))\)

This node is not a leaf node. Assumes binary tree.

updateDirectionalMins :: EncodableStaticCharacter c => SankoffOptimizationDecoration c -> SankoffOptimizationDecoration c -> [StateContributionList] -> SankoffOptimizationDecoration c

Takes two decorations in, a child and a parent, and calculates the median character value of the child. For each possible character state this value is based on whether that character state in the child is on one of the min-cost paths from the root to the leaves. It relies on dynamic programming to do so, using the minimum tuple in the parent to determine whether that character state can participate in the final median. Without loss of generality, for the left child, the character state \(a\) is part of the median if, for some character \(b\) in the child,

\[cost(a_{parent_{leftChild}}) == cost(b_{child}) + transitionCost(a, b).\]

Used on second, pre-order, pass.

calcCostPerState :: Word -> SankoffOptimizationDecoration c -> SankoffOptimizationDecoration c -> ((ExtendedNatural, StateContributionList), (ExtendedNatural, StateContributionList))

Take in a single character state as a Word---which represents an unambiguous character state on the parent--- and two decorations: the decorations of the two child states. Return the minimum costs of transitioning from the input character to each of those two child decorations. These mins will be saved for use at the next post-order call, to the current parent node’s parent.

Note: We can throw away the medians that come back from the tcm here because we’re building medians: the possible character is looped over all available characters, and there's an outer loop which sends in each possible character.