Analysis.Scoring

Description

Decoration

performDecoration :: (DiscreteCharacterMetadata u, DiscreteCharacterMetadata w, DiscreteCharacterDecoration v StaticCharacter, DiscreteCharacterDecoration x StaticCharacter, DiscreteCharacterDecoration y StaticCharacter, RangedCharacterDecoration u ContinuousChar, RangedCharacterDecoration w StaticCharacter, SimpleDynamicDecoration z DynamicChar, Show u, Show v, Show w, Show x, Show y, Show z) => PhylogeneticDAG2 EdgeLength (Maybe String) (Maybe u) (Maybe v) (Maybe w) (Maybe x) (Maybe y) (Maybe z) -> FinalDecorationDAG

Take an undecorated tree and assign peliminary and final states to all nodes.

scoreSolution :: CharacterResult -> PhylogeneticSolution FinalDecorationDAG

Take a solution of one or more undecorated trees and assign peliminary and final states to all nodes.

Decoration Removal

wipeNode

Arguments

:: Monoid n

=> Bool
Do I wipe?

-> PhylogeneticNode2 (CharacterSequence u v w x y z) n

-> PhylogeneticNode2 (CharacterSequence (Maybe u) (Maybe v) (Maybe w) (Maybe x) (Maybe y) (Maybe z)) n

Conditionally wipe the scoring of a single node.

wipeScoring :: Monoid n => PhylogeneticDAG2 e n u v w x y z -> PhylogeneticDAG2 e n (Maybe u) (Maybe v) (Maybe w) (Maybe x) (Maybe y) (Maybe z)

Remove all scoring data from nodes.