Bio.Character

Description

Character Types

data ContinuousChar

Represents a real-valued range with a minimum lower bound of zero and a maximum upper bound of infinity.

data DynamicChar

Represents an encoded dynamic character, consisting of one or more static characters. Dynamic characters treat entire static characters as the character states of the dynamic character. The dynamic character relies on the encoding of the individual static characters to defined the encoding of the entire dynamic character.

type DynamicChars = Vector DynamicChar

A sequence of many dynamic characters. Probably should be asserted as non-empty.

data ExportableCharacterSequence

A structure used for FFI calls.

bufferChunks contains the bit-packed representation of the character sequence.

Constructors

ExportableCharacterSequence

Fields

exportedElementCountSequence :: Word

exportedElementWidthSequence :: Word

exportedBufferChunks :: [CULong]

data StaticCharacter

Represents an encoded static character. Supports binary and numeric operations.

data StaticCharacterBlock

Represents an encoded stream of static characters, consisting of one or more static characters. The static character stream relies on the encoding of the individual static characters to defined the encoding of the entire static character stream.

Character Classes

class ContinuousCharacter c where

A character class for continuous characters.

Minimal complete definition

toContinuousCharacter

Methods

toContinuousCharacter :: Real r => Maybe r -> c

class (EncodableStream s, PossiblyMissingCharacter s) => EncodableDynamicCharacter s where

Represents a character of variable length representing multiple encoded static characters.

decodeMany alphabet . encodeMany alphabet == fmap toList . toList

Minimal complete definition

constructDynamic, encodeDynamic

Methods

constructDynamic :: Foldable t => t (Element s) -> s

Directly construct a dynamic character from the encoded elements.

encodeDynamic :: (Ord a, Foldable t, Foldable c, IsString a) => Alphabet a -> c (t a) -> s

Deprecated: Don't use encodeDynamic, use encodeStream instead!

class (EncodableStreamElement c, PossiblyMissingCharacter c) => EncodableStaticCharacter c where

Represents a single static character encoded in a binary and coercable to Int values.

Binary encoding must allow for the nth symbol in the Alphabet to be tested for membership in the encoded ambiguity group by the following computation:

>>> staticCharacter `testBit` n

Methods

encodeStatic :: (Eq a, IsString a) => Alphabet a -> AmbiguityGroup a -> c

emptyStatic :: c -> c

class (EncodableStaticCharacter (Element s), EncodableStream s, Semigroup s) => EncodableStaticCharacterStream s where

Represents a fixed number of static charcters encoded over Alphabets of the same size and with the same cost structure.

Minimal complete definition

constructStaticStream

Methods

constructStaticStream :: NonEmpty (Element s) -> s

Directly construct a static character stream from the encoded elements.

class (EncodableStreamElement (Element s), MonoFoldable s, MonoFunctor s) => EncodableStream s where

Represents a non empty stream of EncodableStreamElement of variable length.

Laws:

decodeMany alphabet . encodeMany alphabet == fmap toList . toList

Minimal complete definition

encodeStream

Methods

decodeStream :: (Ord a, IsString a) => Alphabet a -> s -> NonEmpty (AmbiguityGroup a)

encodeStream :: (Ord a, IsString a) => Alphabet a -> NonEmpty (AmbiguityGroup a) -> s

indexStream :: s -> Int -> Element s

lookupStream :: s -> Int -> Maybe (Element s)

gapOfStream :: s -> Element s

class (FiniteBits b, EncodedAmbiguityGroupContainer b) => EncodableStreamElement b where

Represents a character of fixed width encoding one or more character states.

Laws:

decodeElement alphabet . encodeChar alphabet . toList == id

encodeElement alphabet [alphabet ! i] == bit i

encodeElement alphabet alphabet == complement (bit (length alphabet - 1) clearBit (bit (length alphabet - 1))

decodeElement alphabet (encodeElement alphabet xs .|. encodeElement alphabet ys) == toList alphabet intersect (toList xs union toList ys)

decodeElement alphabet (encodeElement alphabet xs .&. encodeElement alphabet ys) == toList alphabet intersect (toList xs intersect toList ys)

Minimal complete definition

decodeElement, encodeElement

Methods

decodeElement :: Eq a => Alphabet a -> b -> AmbiguityGroup a

encodeElement :: (Eq a, IsString a) => Alphabet a -> AmbiguityGroup a -> b

getGapElement :: b -> b

getGapChar :: b -> b

Deprecated: Don't use getGapChar, use getGapElement instead!

class Exportable c where

Represents a sequence of fixed width characters packed into a bitwise form consumable by lower level functions.

Minimal complete definition

toExportableBuffer, fromExportableBuffer, toExportableElements, fromExportableElements

Methods

toExportableBuffer :: c -> ExportableCharacterSequence

fromExportableBuffer :: ExportableCharacterSequence -> c

toExportableElements :: c -> Maybe ExportableCharacterElements

fromExportableElements :: ExportableCharacterElements -> c

class PossiblyMissingCharacter c where

Represents a type which may be a missing character. Allows for O(1) testing and construction of missing character.

Minimal complete definition

toMissing, isMissing

Methods

toMissing :: c -> c

isMissing :: c -> Bool

showStream :: EncodableStream s => Alphabet String -> s -> String

Show an EncodableStream by decoding it with it's corresponding alphabet.

showStreamElement :: EncodableStreamElement e => Alphabet String -> e -> String

Show an EncodableStreamElement by decoding it with it's corresponding alphabet.