Bio.Character.Decoration.Additive

Description

Documentation

data AdditivePostorderDecoration a

An abstract initial dynamic character decoration with a polymorphic character type.

data AdditiveOptimizationDecoration a

Represents the finalized character decoration after a pre-order traversal.

class (HasCharacterName s CharacterName, HasCharacterWeight s Double) => GeneralCharacterMetadata s

A constraint for type containing metadata.

Minimal complete definition

extractGeneralCharacterMetadata

class (GeneralCharacterMetadata s, HasCharacterAlphabet s (Alphabet String)) => DiscreteCharacterMetadata s

A decoration of an initial encoding of a dynamic character which has the appropriate Lens & character class constraints.

Minimal complete definition

extractDiscreteCharacterMetadata

class (HasDiscreteCharacter s a, EncodableStaticCharacter a, DiscreteWithTcmCharacterMetadata s a) => DiscreteCharacterDecoration s a | s -> a

(✔)

class HasCharacterAlphabet s a | s -> a where

A Lens for the alphabet field

Minimal complete definition

characterAlphabet

Methods

characterAlphabet :: Lens' s a

class HasCharacterName s a | s -> a where

A Lens for the characterName field

Minimal complete definition

characterName

Methods

characterName :: Lens' s a

class HasSymbolChangeMatrix s a | s -> a where

A Lens for the symbolChangeMatrix field

Minimal complete definition

symbolChangeMatrix

Methods

symbolChangeMatrix :: Lens' s a

class HasTransitionCostMatrix s a | s -> a where

A Lens for the transitionCostMatrix field

Minimal complete definition

transitionCostMatrix

Methods

transitionCostMatrix :: Lens' s a

class HasCharacterWeight s a | s -> a where

A Lens for the characterWeight field

Minimal complete definition

characterWeight

Methods

characterWeight :: Lens' s a

class HasDiscreteCharacter s a | s -> a where

A Lens for the discreteCharacter field

Minimal complete definition

discreteCharacter

Methods

discreteCharacter :: Lens' s a

class HasIntervalCharacter s a | s -> a where

A Lens for the intervalCharacter field.

Minimal complete definition

intervalCharacter

Methods

intervalCharacter :: Lens' s a

class HasIsLeaf s a | s -> a where

A Lens for the isLeaf field.

Minimal complete definition

isLeaf

Methods

isLeaf :: Lens' s a

class HasCharacterCost s a | s -> a where

A Lens for the characterCost field.

Minimal complete definition

characterCost

Methods

characterCost :: Lens' s a

class HasPreliminaryInterval s a | s -> a where

A Lens for the additivePreliminaryInterval field.

Minimal complete definition

preliminaryInterval

Methods

preliminaryInterval :: Lens' s a

class HasFinalInterval s a | s -> a where

A Lens for the additiveFinalInterval field.

Minimal complete definition

finalInterval

Methods

finalInterval :: Lens' s a

class HasChildPrelimIntervals s a | s -> a where

A Lens for the additiveChildPrelimIntervals field.

Minimal complete definition

childPrelimIntervals

Methods

childPrelimIntervals :: Lens' s a

class RangedDecorationOptimization s c => RangedExtensionPreorder s c | s -> c where

A character decoration that can extended to a represent the results of a pre-order traversal.

Minimal complete definition

extendRangedToPreorder

Methods

extendRangedToPreorder :: (DiscreteCharacterMetadata x, RangedPostorderDecoration x c) => x -> Range (Bound c) -> s

class RangedPostorderDecoration s c => RangedExtensionPostorder s c | s -> c where

An abstract initial additive character decoration with a polymorphic character type.

A character decoration that can extended to a represent the results of a post-order traversal.

Minimal complete definition

extendRangedToPostorder

Methods

extendRangedToPostorder

Arguments

:: (DiscreteCharacterMetadata x, RangedCharacterDecoration x c)

=> x
Input decoration

-> Finite (Bound c)
Local cost

-> Range (Bound c)
Preliminary interval

-> (Range (Bound c), Range (Bound c))
Child intervals

-> Bool
Is leaf node?

-> s

class (HasIntervalCharacter d c, Ranged c, ExtendedNumber (Bound c), Num (Finite (Bound c)), Num (Bound c), Ord (Bound c)) => RangedCharacterDecoration d c

Represents a character decoration that has a bounded interval for the character.

Used for Additive character scoring.

class (RangedCharacterDecoration s c, HasFinalInterval s (Range (Bound c))) => RangedDecorationOptimization s c | s -> c

A finalized character decoration that contains an interval character after a pre-order traversal.

class (RangedCharacterDecoration s c, HasCharacterCost s (Finite (Bound c)), HasChildPrelimIntervals s (Range (Bound c), Range (Bound c)), HasIsLeaf s Bool, HasPreliminaryInterval s (Range (Bound c))) => RangedPostorderDecoration s c | s -> c

A partially complete character decoration that contains an interval character after a post-order traversal.