Bio.Character.Decoration.Continuous

Description

Documentation

data ContinuousOptimizationDecoration c

Represents a character decoration after a pre-order traversal.

data ContinuousPostorderDecoration c

Represents the partially complete character decoration after a post-order traversal.

class ContinuousDecoration s a => ContinuousCharacterDecoration s a | s -> a

An abstract initial additive character decoration with a polymorphic character type.

class ContinuousAdditiveHybridDecoration s c => DiscreteExtensionContinuousDecoration s c | s -> c where

A decoration that can be constructed from a DiscreteCharacterDecoration by extending the decoration to contain the requisite fields for performing Continuous's algorithm.

Minimal complete definition

extendDiscreteToContinuous

Methods

extendDiscreteToContinuous :: DiscreteCharacterDecoration x c => x -> Double -> (Double, Double) -> ((Double, Double), (Double, Double)) -> Bool -> s

class (HasCharacterName s CharacterName, HasCharacterWeight s Double) => GeneralCharacterMetadata s

A constraint for type containing metadata.

Minimal complete definition

extractGeneralCharacterMetadata

class (GeneralCharacterMetadata s, HasCharacterAlphabet s (Alphabet String)) => DiscreteCharacterMetadata s

A decoration of an initial encoding of a dynamic character which has the appropriate Lens & character class constraints.

Minimal complete definition

extractDiscreteCharacterMetadata

class (HasDiscreteCharacter s a, EncodableStaticCharacter a, DiscreteWithTcmCharacterMetadata s a) => DiscreteCharacterDecoration s a | s -> a

(✔)

class HasCharacterName s a | s -> a where

A Lens for the characterName field

Minimal complete definition

characterName

Methods

characterName :: Lens' s a

class HasCharacterWeight s a | s -> a where

A Lens for the characterWeight field

Minimal complete definition

characterWeight

Methods

characterWeight :: Lens' s a

class HasIsLeaf s a | s -> a where

A Lens for the isLeaf field.

Minimal complete definition

isLeaf

Methods

isLeaf :: Lens' s a

class HasCharacterCost s a | s -> a where

A Lens for the characterCost field.

Minimal complete definition

characterCost

Methods

characterCost :: Lens' s a

class HasPreliminaryInterval s a | s -> a where

A Lens for the additivePreliminaryInterval field.

Minimal complete definition

preliminaryInterval

Methods

preliminaryInterval :: Lens' s a

class HasChildPrelimIntervals s a | s -> a where

A Lens for the additiveChildPrelimIntervals field.

Minimal complete definition

childPrelimIntervals

Methods

childPrelimIntervals :: Lens' s a

class HasIntervalCharacter s a | s -> a where

A Lens for the intervalCharacter field.

Minimal complete definition

intervalCharacter

Methods

intervalCharacter :: Lens' s a

continuousDecorationInitial :: CharacterName -> Double -> c -> ContinuousDecorationInitial c

A smart constructor for a continuous character.

data ContinuousDecorationInitial c

An abstract initial continuous character decoration with a polymorphic character type.

class HasContinuousCharacter s a | s -> a where

A Lens for the continuousCharacter field

Minimal complete definition

continuousCharacter

Methods

continuousCharacter :: Lens' s a