Bio.Character.Decoration.Discrete

Description

Documentation

data DiscreteDecoration c

General, concrete type for Discrete characters.

class (HasDiscreteCharacter s a, EncodableStaticCharacter a, DiscreteWithTcmCharacterMetadata s a) => DiscreteCharacterDecoration s a | s -> a

(✔)

class (GeneralCharacterMetadata s, HasCharacterAlphabet s (Alphabet String)) => DiscreteCharacterMetadata s

A decoration of an initial encoding of a dynamic character which has the appropriate Lens & character class constraints.

Minimal complete definition

extractDiscreteCharacterMetadata

class (HasCharacterName s CharacterName, HasCharacterWeight s Double) => GeneralCharacterMetadata s

A constraint for type containing metadata.

Minimal complete definition

extractGeneralCharacterMetadata

class HasCharacterAlphabet s a | s -> a where

A Lens for the alphabet field

Minimal complete definition

characterAlphabet

Methods

characterAlphabet :: Lens' s a

class HasCharacterName s a | s -> a where

A Lens for the characterName field

Minimal complete definition

characterName

Methods

characterName :: Lens' s a

class HasCharacterWeight s a | s -> a where

A Lens for the characterWeight field

Minimal complete definition

characterWeight

Methods

characterWeight :: Lens' s a

class HasDiscreteCharacter s a | s -> a where

A Lens for the discreteCharacter field

Minimal complete definition

discreteCharacter

Methods

discreteCharacter :: Lens' s a

class HasSymbolChangeMatrix s a | s -> a where

A Lens for the symbolChangeMatrix field

Minimal complete definition

symbolChangeMatrix

Methods

symbolChangeMatrix :: Lens' s a

class HasTransitionCostMatrix s a | s -> a where

A Lens for the transitionCostMatrix field

Minimal complete definition

transitionCostMatrix

Methods

transitionCostMatrix :: Lens' s a

class PossiblyMissingCharacter c where

Represents a type which may be a missing character. Allows for O(1) testing and construction of missing character.

Minimal complete definition

toMissing, isMissing

Methods

toMissing :: c -> c

isMissing :: c -> Bool

class DiscreteCharacterDecoration s a => SimpleDiscreteCharacterDecoration s a | s -> a where

(✔)

Minimal complete definition

toDiscreteCharacterDecoration

Methods

toDiscreteCharacterDecoration :: CharacterName -> Double -> Alphabet String -> (Word -> Word -> Word) -> (x -> a) -> x -> s

showDiscreteCharacterElement :: (EncodableStreamElement a, HasCharacterAlphabet s (Alphabet String), HasDiscreteCharacter s a) => s -> String

Show an appropriate instance of HasDiscreteCharacter that is also HasCharacterAlphabet by decoding the EncodableStreamElement over the Alphabet; both of which contained in the lensed type.

Orphan instances

(HasDiscreteCharacter s c, PossiblyMissingCharacter c) => PossiblyMissingCharacter s
(✔)

Methods

toMissing :: s -> s

isMissing :: s -> Bool