Bio.Character.Decoration.Dynamic

Description

Documentation

data DynamicDecorationInitial d

An abstract initial dynamic character decoration with a polymorphic character type.

class (HasCharacterCost s Word, HasCharacterLocalCost s Word, HasPreliminaryGapped s a, HasPreliminaryUngapped s a, HasLeftAlignment s a, HasRightAlignment s a, SimpleDynamicDecoration s a) => DirectOptimizationPostOrderDecoration s a | s -> a

An incomplete decoration of a dynamic character with half the direct optimization annotations.

Represents the result of just the post-order traversal.

Is a sub-class of DynamicCharacterDecoration.

data DynamicDecorationDirectOptimization d

An abstract direct optimization dynamic character decoration with a polymorphic character type.

data DynamicDecorationImpliedAlignment d

An abstract implied alignment dynamic character decoration with a polymorphic character type.

class (HasAverageLength s AverageLength, HasEncoded s a, EncodableDynamicCharacter a, DynamicCharacterMetadata s (Element a)) => SimpleDynamicDecoration s a | s -> a

A decoration of an initial encoding of a dynamic character which has the appropriate Lens & character class constraints.

class (HasFinalGapped s a, HasFinalUngapped s a, DirectOptimizationPostOrderDecoration s a) => DirectOptimizationDecoration s a | s -> a

A decoration of a dynamic character with all direct optimization annotations.

Is a sub-class of DirectOptimizationPreOrderDecoration.

class (HasImpliedAlignment s a, DirectOptimizationDecoration s a) => ImpliedAlignmentDecoration s a | s -> a

A decoration of a dynamic character with the implied alignment decorations.

Is a sub-class of DirectOptimizationDecoration.

class SimpleDynamicDecoration s a => DynamicCharacterDecoration s a | s -> a where

A decoration of an initial encoding of a dynamic character which has the appropriate Lens & character class constraints.

Minimal complete definition

toDynamicCharacterDecoration

Methods

toDynamicCharacterDecoration :: CharacterName -> Double -> Alphabet String -> (Word -> Word -> Word) -> (x -> a) -> x -> s

data DynamicDecorationDirectOptimizationPostOrderResult d

Represents the partial character decoration result of a post-order traversal.

class (HasCharacterName s CharacterName, HasCharacterWeight s Double) => GeneralCharacterMetadata s

A constraint for type containing metadata.

Minimal complete definition

extractGeneralCharacterMetadata

class (GeneralCharacterMetadata s, HasCharacterAlphabet s (Alphabet String)) => DiscreteCharacterMetadata s

A decoration of an initial encoding of a dynamic character which has the appropriate Lens & character class constraints.

Minimal complete definition

extractDiscreteCharacterMetadata

class (DiscreteCharacterMetadata s, EncodableStreamElement c, HasSymbolChangeMatrix s (Word -> Word -> Word), HasTransitionCostMatrix s (c -> c -> (c, Word))) => DiscreteWithTcmCharacterMetadata s c | s -> c

A decoration of an initial encoding of a dynamic character which has the appropriate Lens & character class constraints.

class (DiscreteWithTcmCharacterMetadata s c, HasDenseTransitionCostMatrix s (Maybe DenseTransitionCostMatrix), HasSparseTransitionCostMatrix s MemoizedCostMatrix, HasTraversalFoci s (Maybe TraversalFoci)) => DynamicCharacterMetadata s c | s -> c

A decoration of an initial encoding of a dynamic character which has the appropriate Lens & character class constraints.

Minimal complete definition

extractDynamicCharacterMetadata

class (SimpleDynamicDecoration s c, DirectOptimizationPostOrderDecoration s c) => SimpleDynamicExtensionPostOrderDecoration s c | s -> c where

A decoration that can be constructed from a DiscreteCharacterDecoration by extending the decoration to contain the requisite fields for performing Sankoff's algorithm.

Minimal complete definition

extendDynamicToPostOrder

Methods

extendDynamicToPostOrder

Arguments

:: SimpleDynamicDecoration x c

=> x
Original decoration

-> Word
The cost of the alignment

-> Word
The cost of the alignment and the child subtrees

-> AverageLength
The average length of the dynamic character in the subtree

-> c
Preliminary ungapped dynamic character

-> c
Preliminary gapped dynamic character

-> c
Left alignment dynamic character

-> c
Right alignment dynamic character

-> s
Resulting decoration

class (DirectOptimizationPostOrderDecoration s c, DirectOptimizationDecoration s c) => PostOrderExtensionDirectOptimizationDecoration s c | s -> c where

A decoration that can be constructed from a DiscreteCharacterDecoration by extending the decoration to contain the requisite fields for performing Sankoff's algorithm.

Minimal complete definition

extendPostOrderToDirectOptimization

Methods

extendPostOrderToDirectOptimization

Arguments

:: DirectOptimizationPostOrderDecoration x c

=> x
Original decoration

-> c
Final ungapped dynamic character

-> c
Final gapped dynamic character

-> c
Final single dynamic character

-> s
Resulting decoration

class HasAverageLength s a | s -> a where

A Lens for the impliedAlignment field

Minimal complete definition

averageLength

Methods

averageLength :: Lens' s a

class HasCharacterAlphabet s a | s -> a where

A Lens for the alphabet field

Minimal complete definition

characterAlphabet

Methods

characterAlphabet :: Lens' s a

class HasCharacterCost s a | s -> a where

A Lens for the characterCost field.

Minimal complete definition

characterCost

Methods

characterCost :: Lens' s a

class HasCharacterLocalCost s a | s -> a where

A Lens for the encoded field

Minimal complete definition

characterLocalCost

Methods

characterLocalCost :: Lens' s a

class HasCharacterName s a | s -> a where

A Lens for the characterName field

Minimal complete definition

characterName

Methods

characterName :: Lens' s a

class HasCharacterWeight s a | s -> a where

A Lens for the characterWeight field

Minimal complete definition

characterWeight

Methods

characterWeight :: Lens' s a

class HasEncoded s a | s -> a where

A Lens for the encoded field

Minimal complete definition

encoded

Methods

encoded :: Lens' s a

class HasFinalGapped s a | s -> a where

A Lens for the finalGapped field

Minimal complete definition

finalGapped

Methods

finalGapped :: Lens' s a

class HasFinalUngapped s a | s -> a where

A Lens for the finalUngapped field

Minimal complete definition

finalUngapped

Methods

finalUngapped :: Lens' s a

class HasPreliminaryGapped s a | s -> a where

A Lens for the preliminaryGapped field

Minimal complete definition

preliminaryGapped

Methods

preliminaryGapped :: Lens' s a

class HasPreliminaryUngapped s a | s -> a where

A Lens for the preliminaryUngapped field

Minimal complete definition

preliminaryUngapped

Methods

preliminaryUngapped :: Lens' s a

class HasLeftAlignment s a | s -> a where

A Lens for the leftAlignment field

Minimal complete definition

leftAlignment

Methods

leftAlignment :: Lens' s a

class HasRightAlignment s a | s -> a where

A Lens for the rightAlignment field

Minimal complete definition

rightAlignment

Methods

rightAlignment :: Lens' s a

class HasImpliedAlignment s a | s -> a where

A Lens for the impliedAlignment field

Minimal complete definition

impliedAlignment

Methods

impliedAlignment :: Lens' s a

class HasSingleDisambiguation s a | s -> a where

A Lens for the singleDisambiguation field

Minimal complete definition

singleDisambiguation

Methods

singleDisambiguation :: Lens' s a

class HasSymbolChangeMatrix s a | s -> a where

A Lens for the symbolChangeMatrix field

Minimal complete definition

symbolChangeMatrix

Methods

symbolChangeMatrix :: Lens' s a

class HasTransitionCostMatrix s a | s -> a where

A Lens for the transitionCostMatrix field

Minimal complete definition

transitionCostMatrix

Methods

transitionCostMatrix :: Lens' s a

class HasTraversalFoci s a | s -> a where

A Lens for the traversalFoci field

Minimal complete definition

traversalFoci

Methods

traversalFoci :: Lens' s a

class HasSparseTransitionCostMatrix s a | s -> a where

A Lens for the denseTransitionCostMatrix field

Minimal complete definition

sparseTransitionCostMatrix

Methods

sparseTransitionCostMatrix :: Lens' s a

class HasDenseTransitionCostMatrix s a | s -> a where

A Lens for the denseTransitionCostMatrix field

Minimal complete definition

denseTransitionCostMatrix

Methods

denseTransitionCostMatrix :: Lens' s a

class PossiblyMissingCharacter c where

Represents a type which may be a missing character. Allows for O(1) testing and construction of missing character.

Minimal complete definition

toMissing, isMissing

Methods

toMissing :: c -> c

isMissing :: c -> Bool

data AverageLength

The average length of all the sequences in the subtree.

Forms a Semigroup for effcient, recursive post-order acumulation on the tree.

toAverageLength :: Word -> AverageLength

Safely construct an AverageLength from a non-negative value.

getAverageLength :: Fractional a => AverageLength -> a

Safely convert an AverageLength to a Fractional representation.

type TraversalFoci = NonEmpty TraversalFocus

Represents a collection of paired rooting edges and unrooted topologies.

type TraversalFocusEdge = (Int, Int)

An unrooted edge.

type TraversalTopology = TopologyRepresentation TraversalFocusEdge

A unique representation of a DAG topology.