Bio.Character.Decoration.Fitch

Description

Documentation

class DiscreteCharacterDecoration s a => FitchCharacterDecoration s a | s -> a

An optimization class for Fitch (non-additive) characters.

data FitchOptimizationDecoration f

An abstract initial dynamic character decoration with a polymorphic character type.

Constructors

FitchOptimizationDecoration Word f f (f, f) Bool f DiscreteCharacterMetadataDec

class FitchDecoration s c => DiscreteExtensionFitchDecoration s c | s -> c where

A decoration that can be constructed from a DiscreteCharacterDecoration by extending the decoration to contain the requisite fields for performing Fitch's algorithm.

Minimal complete definition

extendDiscreteToFitch

Methods

extendDiscreteToFitch :: DiscreteCharacterDecoration x c => x -> Word -> c -> c -> (c, c) -> Bool -> s

class (HasCharacterName s CharacterName, HasCharacterWeight s Double) => GeneralCharacterMetadata s

A constraint for type containing metadata.

Minimal complete definition

extractGeneralCharacterMetadata

class (GeneralCharacterMetadata s, HasCharacterAlphabet s (Alphabet String)) => DiscreteCharacterMetadata s

A decoration of an initial encoding of a dynamic character which has the appropriate Lens & character class constraints.

Minimal complete definition

extractDiscreteCharacterMetadata

class (HasDiscreteCharacter s a, EncodableStaticCharacter a, DiscreteWithTcmCharacterMetadata s a) => DiscreteCharacterDecoration s a | s -> a

(✔)

class HasCharacterAlphabet s a | s -> a where

A Lens for the alphabet field

Minimal complete definition

characterAlphabet

Methods

characterAlphabet :: Lens' s a

class HasCharacterName s a | s -> a where

A Lens for the characterName field

Minimal complete definition

characterName

Methods

characterName :: Lens' s a

class HasCharacterWeight s a | s -> a where

A Lens for the characterWeight field

Minimal complete definition

characterWeight

Methods

characterWeight :: Lens' s a

class HasDiscreteCharacter s a | s -> a where

A Lens for the discreteCharacter field

Minimal complete definition

discreteCharacter

Methods

discreteCharacter :: Lens' s a

class HasFinalMedian s a | s -> a where

A Lens for the HasFinalMedian field.

Minimal complete definition

finalMedian

Methods

finalMedian :: Lens' s a

class HasSymbolChangeMatrix s a | s -> a where

A Lens for the symbolChangeMatrix field

Minimal complete definition

symbolChangeMatrix

Methods

symbolChangeMatrix :: Lens' s a

class HasTransitionCostMatrix s a | s -> a where

A Lens for the transitionCostMatrix field

Minimal complete definition

transitionCostMatrix

Methods

transitionCostMatrix :: Lens' s a

class HasIsLeaf s a | s -> a where

A Lens for the isLeaf field.

Minimal complete definition

isLeaf

Methods

isLeaf :: Lens' s a

class HasCharacterCost s a | s -> a where

A Lens for the characterCost field.

Minimal complete definition

characterCost

Methods

characterCost :: Lens' s a

class HasPreliminaryMedian s a | s -> a where

A Lens for the fitchPreliminaryMedian field.

Minimal complete definition

preliminaryMedian

Methods

preliminaryMedian :: Lens' s a

class HasChildMedians s a | s -> a where

A Lens for the fitchChildMedians field.

Minimal complete definition

childMedians

Methods

childMedians :: Lens' s a