Bio.Character.Decoration.Metric

Description

Documentation

data MetricDecorationInitial c

An abstract initial dynamic character decoration with a polymorphic character type.

class DiscreteCharacterDecoration s a => MetricCharacterDecoration s a | s -> a

A character optimization class for TCM which satisfy the properties of metricity.

class SankoffDecoration s c => DiscreteExtensionSankoffDecoration s c | s -> c where

A decoration that can be constructed from a DiscreteCharacterDecoration by extending the decoration to contain the requisite fields for performing Sankoff's algorithm.

Minimal complete definition

extendDiscreteToSankoff

Methods

extendDiscreteToSankoff :: DiscreteCharacterDecoration x c => x -> [ExtendedNatural] -> [ExtendedNatural] -> [ExtendedNatural] -> [ExtendedNatural] -> ([[Word]], [[Word]]) -> Word -> c -> Bool -> s

class (GeneralCharacterMetadata s, HasCharacterAlphabet s (Alphabet String)) => DiscreteCharacterMetadata s

A decoration of an initial encoding of a dynamic character which has the appropriate Lens & character class constraints.

Minimal complete definition

extractDiscreteCharacterMetadata

class (HasDiscreteCharacter s a, EncodableStaticCharacter a, DiscreteWithTcmCharacterMetadata s a) => DiscreteCharacterDecoration s a | s -> a

(✔)

class (HasCharacterName s CharacterName, HasCharacterWeight s Double) => GeneralCharacterMetadata s

A constraint for type containing metadata.

Minimal complete definition

extractGeneralCharacterMetadata

class HasBeta s a | s -> a where

A Lens for the beta field.

Minimal complete definition

beta

Methods

beta :: Lens' s a

class HasCharacterAlphabet s a | s -> a where

A Lens for the alphabet field

Minimal complete definition

characterAlphabet

Methods

characterAlphabet :: Lens' s a

class HasCharacterCost s a | s -> a where

A Lens for the characterCost field.

Minimal complete definition

characterCost

Methods

characterCost :: Lens' s a

class HasCharacterCostVector s a | s -> a where

A Lens for the minCostVector field.

Minimal complete definition

characterCostVector

Methods

characterCostVector :: Lens' s a

class HasCharacterName s a | s -> a where

A Lens for the characterName field

Minimal complete definition

characterName

Methods

characterName :: Lens' s a

class HasCharacterWeight s a | s -> a where

A Lens for the characterWeight field

Minimal complete definition

characterWeight

Methods

characterWeight :: Lens' s a

class HasDiscreteCharacter s a | s -> a where

A Lens for the discreteCharacter field

Minimal complete definition

discreteCharacter

Methods

discreteCharacter :: Lens' s a

class HasFinalExtraCost s a | s -> a where

A Lens for the finalExtraCost field.

Minimal complete definition

finalExtraCost

Methods

finalExtraCost :: Lens' s a

class HasIsLeaf s a | s -> a where

A Lens for the isLeaf field.

Minimal complete definition

isLeaf

Methods

isLeaf :: Lens' s a

class HasPreliminaryExtraCost s a | s -> a where

A Lens for the preliminaryExtraCost field.

Minimal complete definition

preliminaryExtraCost

Methods

preliminaryExtraCost :: Lens' s a

class HasStateMinTuple s a | s -> a where

A Lens for the minStateTuple field.

Minimal complete definition

minStateTuple

Methods

minStateTuple :: Lens' s a

class HasSymbolChangeMatrix s a | s -> a where

A Lens for the symbolChangeMatrix field

Minimal complete definition

symbolChangeMatrix

Methods

symbolChangeMatrix :: Lens' s a

class HasTransitionCostMatrix s a | s -> a where

A Lens for the transitionCostMatrix field

Minimal complete definition

transitionCostMatrix

Methods

transitionCostMatrix :: Lens' s a

class (DiscreteCharacterDecoration s c, HasCharacterCost s Word, HasCharacterCostVector s [ExtendedNatural], HasStateMinTuple s ([[Word]], [[Word]])) => SankoffDecoration s c | s -> c

A decoration containing a character that has been scored using Sankoff's algorithm.

data SankoffOptimizationDecoration c

A concrete type representing the results of performing Sankoff's algorithm.

type StateContributionList = [Word]

A list of states on the child that contribute to the lowest score on each state in the parent Used to simplify? SankoffOptimizationDecoration