Bio.Character.Encodable.Continuous

Description

Documentation

data ContinuousChar

Represents a real-valued range with a minimum lower bound of zero and a maximum upper bound of infinity.

class ContinuousCharacter c where

A character class for continuous characters.

Minimal complete definition

toContinuousCharacter

Methods

toContinuousCharacter :: Real r => Maybe r -> c

class PossiblyMissingCharacter c where

Represents a type which may be a missing character. Allows for O(1) testing and construction of missing character.

Minimal complete definition

toMissing, isMissing

Methods

toMissing :: c -> c

isMissing :: c -> Bool