Bio.Character.Exportable

Description

Class for needed operations of coded sequences and characters

Documentation

class Exportable c where

Represents a sequence of fixed width characters packed into a bitwise form consumable by lower level functions.

Minimal complete definition

toExportableBuffer, fromExportableBuffer, toExportableElements, fromExportableElements

Methods

toExportableBuffer :: c -> ExportableCharacterSequence

fromExportableBuffer :: ExportableCharacterSequence -> c

toExportableElements :: c -> Maybe ExportableCharacterElements

fromExportableElements :: ExportableCharacterElements -> c

class HasExportedElementCount s a | s -> a where

A Lens for the exportedElementCount field

Minimal complete definition

exportedElementCount

Methods

exportedElementCount :: Lens' s a

class HasExportedElementWidth s a | s -> a where

A Lens for the exportedElementWidth field

Minimal complete definition

exportedElementWidth

Methods

exportedElementWidth :: Lens' s a

data ExportableCharacterElements

A structure used for FFI calls-- characterElements contains the integral value for each character element.

Constructors

ExportableCharacterElements Word Word [CUInt]

data ExportableCharacterSequence

A structure used for FFI calls.

bufferChunks contains the bit-packed representation of the character sequence.

Constructors

ExportableCharacterSequence Word Word [CULong]