Bio.Character.Parsed

Description

Export of parsed character

Documentation

type Identifier = String

The string value that uniquely identifies a taxon.

data ParsedCharacter

A generalized character type extracted from a parser. A character can be real-valued, discrete and singular, or discrete with variable length.

Constructors

ParsedContinuousCharacter (Maybe Double)

ParsedDiscreteCharacter (Maybe (AmbiguityGroup String))

ParsedDynamicCharacter (Maybe (NonEmpty (AmbiguityGroup String)))

type ParsedChars = Vector ParsedCharacter

Represents a character sequence containing possibly missing character data.

class ParsedCharacters a where

Instances provide a method to extract Character sequences from raw parsed results. The TreeSeqs are agnostic of character data types. Tree-only return values from files will extract the taxa labels from leaf nodes only with empty sequences.

Characters of types DNA, RNA, protein, and amino acid will *not* have thier IUPAC codes translated to the apropriate groups. This abiguity group translation will occur later during the rectification process with the character metadata. Parsers which produce expanded ambiguity groups for these character types will be collapsed back to the IUPAC code for the ambiguity group during the type-class's extraction process.

It is expected that parsers will altered to return simpler character literals for time efficientcy in the future.

I need to think about how this might interact with some things in Nexus, but it seems to make sense. It might make verification in the parsers more difficult... thinking...

Minimal complete definition

unifyCharacters

Methods

unifyCharacters :: a -> TaxonCharacters

type TaxonCharacters = Map Identifier ParsedChars

A mapping from taxon identifiers to thier corresponding sequences.