Bio.Graph.Component

Description

Documentation

class PhylogeneticComponent t i e n | t -> i, t -> n, t -> e where

Represents the most relaxed phylogentic graph structure.

The graph must satisfy the following: * The graph is directed * The graph is acyclic * the graph contains one or more root nodes * All nodes have at most in-degree 2 * All nodes have out-degree 0 or out-degree 2

Minimal complete definition

parents, children, roots, leaves, nodeCount, nodeDatum, edgeDatum, isComponentNode, isNetworkNode, isTreeNode, isLeafNode, isRootNode, networkResolutions

Methods

parents :: i -> t -> [i]

children :: i -> t -> [i]

roots :: t -> NonEmpty i

leaves :: t -> [i]

nodeCount :: t -> Int

nodeDatum :: i -> t -> n

edgeDatum :: (i, i) -> t -> Maybe e

isComponentNode :: i -> t -> Bool

A node satisfying: * In-degree of 2 or more * Out-degree of 2 or more * A least 2 parent nodes have ancestoral paths to different root nodes

isNetworkNode :: i -> t -> Bool

A node satisfying: * In-degree of 2 or more * Out-degree of 2 or more * All parent nodes have ancestoral paths to a single root node

isTreeNode :: i -> t -> Bool

A node satisfying: * In-degree of 1 * Out-degree of 2

isLeafNode :: i -> t -> Bool

A node satisfying: * Out-degree 0

isRootNode :: i -> t -> Bool

A node satisfying: * In-degree 0

networkResolutions :: t -> NonEmpty t

Performs a softwire resolution of all component nodes into a collection of all resulting networks. The resulting size of the collection is equal to 2^n where n is the number of component nodes in the PhylogeneticComponent.

class PhylogeneticNetwork t i e n | t -> i, t -> n, t -> e where

Represents a more constrained phylogentic graph structure.

The graph must satisfy the following: * The graph is directed * The graph is acyclic * The graph contains exactly one root nodes * All nodes have at most in-degree 2 * All nodes have out-degree 0 or out-degree 2

Minimal complete definition

root, treeResolutions

Methods

root :: t -> i

treeResolutions :: t -> NonEmpty t

Performs a softwire resolution of all network nodes into a collection of all resulting trees. The resulting size of the collection is equal to 2^n where n is the number of network nodes in the PhylogenetiNetwork.

class PhylogeneticTree t i e n | t -> i, t -> n, t -> e where

Represents the most constrained phylogentic graph structure. The constraints correlate to a binary tree.

The graph must satisfy the following: * The graph is directed * The graph is acyclic * The graph contains exactly one root nodes * All nodes have exactly in-degree 1 * All nodes contain out-degree 0 or out-degree 2

Minimal complete definition

parent

Methods

parent :: i -> t -> Maybe i