Bio.Graph.Forest.Parsed

Description

Typeclass for a parsed forest so that it can convert into an internal forest.

Documentation

type ParserTree = ReferenceDAG () EdgeLength (Maybe String)

The type of possibly present decorations on a tree from a parsed file.

type ParserForest = Maybe (PhylogeneticForest ParserTree)

The parser coalesced type, representing a possibly present forest.

type ParserForestSet = Maybe (NonEmpty (PhylogeneticForest ParserTree))

The parser coalesced type, representing a possibly present forest.

data NewickEnum

An internal type for representing a node with a unique numeric identifier.

Constructors

NE !Int (Maybe String) (Maybe Double) [NewickEnum]

class ParsedForest a where

Represents a parser result type which can have a possibly empty forest extracted from it.

Minimal complete definition

unifyGraph

Methods

unifyGraph :: a -> ParserForestSet

Orphan instances

Hashable GraphID
(✔)

Methods

hashWithSalt :: Int -> GraphID -> Int

hash :: GraphID -> Int