Bio.Graph.Node

Description

Documentation

data EdgeSet e

Represents a collection of edges.

Often used to represent a spanning tree in a DAG.

data NewickSerialization

A newick representation of a subtree. Semigroup instance used for subtree joining.

data PhylogeneticNode n s

This serves as a computation invariant node decoration designed to hold node information such as name and later a subtree structure.

Constructors

PNode

Fields

nodeDecorationDatum :: n

sequenceDecoration :: s

data PhylogeneticNode2 s n

This serves as a computation dependant node decoration designed to hold node information for a a phylogenetic network (or tree).

Constructors

PNode2

Fields

resolutions :: ResolutionCache s

nodeDecorationDatum2 :: n

type ResolutionCache s = NonEmpty (ResolutionInformation s)

A collection of subtree resolutions. Represents a non-deterministic collection of subtree choices.

data ResolutionInformation s

A collection of information used to memoize network optimizations.

Constructors

ResInfo

Fields

totalSubtreeCost :: Double

localSequenceCost :: Double

leafSetRepresentation :: SubtreeLeafSet

subtreeRepresentation :: NewickSerialization

subtreeEdgeSet :: EdgeSet (Int, Int)

topologyRepresentation :: TopologyRepresentation (Int, Int)

characterSequence :: s

data SubtreeLeafSet

An arbitrarily ordered collection of leaf nodes in a subtree. Leaves in the tree are uniquely identified by an index across the entire DAG. Set bits represent a leaf uniquily identified by that index being present in the subtree.

Use the Semigroup operation '(<>)' to union the leaves included in a leaf set.

addEdgeToEdgeSet :: (Int, Int) -> ResolutionInformation s -> ResolutionInformation s

Adds an edge reference to an existing subtree resolution.

addNetworkEdgeToTopology

Arguments

:: (Int, Int)
Applied network edge identifier

-> (Int, Int)
Excluded network edge identifier

-> ResolutionInformation s

-> ResolutionInformation s

Updates the TopologyRepresentation to include a new network edge present in to spanning tree the node is a subtree of.

singletonEdgeSet :: e -> EdgeSet e

Construct a singleton EdgeSet value. Use the semigroup operator '(<>)' to construct larger a EdgeSet. This enforces the non-empty invariant of the EdgeSet data-structure.

singletonNewickSerialization :: Show i => i -> NewickSerialization

Construct a singleton newick string with a unique identifier that can be rendered to a string through it's Show instance.

singletonSubtreeLeafSet

Arguments

:: Int
Leaf count

-> Int
Leaf index

-> SubtreeLeafSet

Construct a singleton leaf set by supplying the number of leaves and the unique leaf index.

pNode2 :: n -> ResolutionCache s -> PhylogeneticNode2 s n

A safe constructor of a PhylogeneticNode2.