Bio.Graph.PhylogeneticDAG

Description

Containing the master command for unifying all input types: tree, metadata, and sequence

Documentation

newtype PhylogeneticDAG e n u v w x y z

Wrapper for ReferenceDAG (deprecated)

Type annotations (metadata types):

e = EdgeLength

n = node labels: Maybe(String)

u = various (initial, post-order, pre-order) Continuous specified as ContinuousChar or General

v = various (initial, post-order, pre-order) Fitch specified as StaticCharacter or Discrete

w = various (initial, post-order, pre-order) Additive specified as StaticCharacter or Discrete

x = various (initial, post-order, pre-order) Sankoff specified as StaticCharacter or Discrete

y = various (initial, post-order, pre-order) Sankoff specified as StaticCharacter or Discrete

z = various (initial, post-order, pre-order) Dynamic specified as DynamicChar or DiscreteWithTCM

Constructors

PDAG (ReferenceDAG () e (PhylogeneticNode n (CharacterSequence u v w x y z)))

newtype PhylogeneticDAG2 e n u v w x y z

Wrapper for ReferenceDAG Type annotations (metadata types):

e = edge info, as yet undetermined

n = node labels: Maybe(String)

u = various (initial, post-order, pre-order) Continuous specified as ContinuousChar or General

v = various (initial, post-order, pre-order) Fitch specified as StaticCharacter or Discrete

w = various (initial, post-order, pre-order) Additive specified as StaticCharacter or Discrete

x = various (initial, post-order, pre-order) Sankoff specified as StaticCharacter or Discrete

y = various (initial, post-order, pre-order) Sankoff specified as StaticCharacter or Discrete

z = various (initial, post-order, pre-order) Dynamic specified as DynamicChar or DiscreteWithTCM

Constructors

PDAG2 (ReferenceDAG (HashMap EdgeReference (ResolutionCache (CharacterSequence u v w x y z)), Vector (HashMap EdgeReference (ResolutionCache (CharacterSequence u v w x y z))), Maybe (NonEmpty (TraversalTopology, Double, Double, Double, Vector (NonEmpty TraversalFocusEdge)))) e (PhylogeneticNode2 (CharacterSequence u v w x y z) n))

class PhylogeneticDAGish dag where

A typeclass for abstracting opperations on a phylogenetic DAG.

Methods

postorderDAG :: (e -> [e'] -> e') -> (n -> [n'] -> n') -> (u -> [u'] -> u') -> (v -> [v'] -> v') -> (w -> [w'] -> w') -> (x -> [x'] -> x') -> (y -> [y'] -> y') -> (z -> [z'] -> z') -> dag e n u v w x y z -> dag e' n' u' v' w' x' y' z'

postorderSequence :: (u -> [u'] -> u') -> (v -> [v'] -> v') -> (w -> [w'] -> w') -> (x -> [x'] -> x') -> (y -> [y'] -> y') -> (z -> [z'] -> z') -> dag e n u v w x y z -> dag e n u' v' w' x' y' z'

postorderEdge :: (e -> [e'] -> e') -> dag e n u v w x y z -> dag e' n u v w x y z

postorderNode :: (n -> [n'] -> n') -> dag e n u v w x y z -> dag e n' u v w x y z

postorderContinuous :: (u -> [u'] -> u') -> dag e n u v w x y z -> dag e n u' v w x y z

postorderNonAdditive :: (v -> [v'] -> v') -> dag e n u v w x y z -> dag e n u v' w x y z

postorderAdditive :: (w -> [w'] -> w') -> dag e n u v w x y z -> dag e n u v w' x y z

postorderMetric :: (x -> [x'] -> x') -> dag e n u v w x y z -> dag e n u v w x' y z

postorderNonMetric :: (y -> [y'] -> y') -> dag e n u v w x y z -> dag e n u v w x y' z

postorderDynamic :: (z -> [z'] -> z') -> dag e n u v w x y z -> dag e n u v w x y z'

preorderDAG :: (e -> [(Word, e')] -> e') -> (n -> [(Word, n')] -> n') -> (u -> [(Word, u')] -> u') -> (v -> [(Word, v')] -> v') -> (w -> [(Word, w')] -> w') -> (x -> [(Word, x')] -> x') -> (y -> [(Word, y')] -> y') -> (z -> [(Word, z')] -> z') -> dag e n u v w x y z -> dag e' n' u' v' w' x' y' z'

preorderSequence :: (u -> [(Word, u')] -> u') -> (v -> [(Word, v')] -> v') -> (w -> [(Word, w')] -> w') -> (x -> [(Word, x')] -> x') -> (y -> [(Word, y')] -> y') -> (z -> [(Word, z')] -> z') -> dag e n u v w x y z -> dag e n u' v' w' x' y' z'

preorderEdge :: (e -> [(Word, e')] -> e') -> dag e n u v w x y z -> dag e' n u v w x y z

preorderNode :: (n -> [(Word, n')] -> n') -> dag e n u v w x y z -> dag e n' u v w x y z

preorderContinuous :: (u -> [(Word, u')] -> u') -> dag e n u v w x y z -> dag e n u' v w x y z

preorderNonAdditive :: (v -> [(Word, v')] -> v') -> dag e n u v w x y z -> dag e n u v' w x y z

preorderAdditive :: (w -> [(Word, w')] -> w') -> dag e n u v w x y z -> dag e n u v w' x y z

preorderMetric :: (x -> [(Word, x')] -> x') -> dag e n u v w x y z -> dag e n u v w x' y z

preorderNonMetric :: (y -> [(Word, y')] -> y') -> dag e n u v w x y z -> dag e n u v w x y' z

preorderDynamic :: (z -> [(Word, z')] -> z') -> dag e n u v w x y z -> dag e n u v w x y z'

assignOptimalDynamicCharacterRootEdges

Arguments

:: (HasBlockCost u v w x y z Word Double, HasTraversalFoci z (Maybe TraversalFoci))

=> (z -> [z] -> z)
Post-order traversal function for Dynamic Characters.

-> PhylogeneticDAG2 e n u v w x y z

-> (PhylogeneticDAG2 e n u v w x y z, HashMap EdgeReference (ResolutionCache (CharacterSequence u v w x y z)), Vector (HashMap EdgeReference (ResolutionCache (CharacterSequence u v w x y z))))

For every edge in the component:

If the edge is *not* a network edge:

The re-rooting candidate cost for that edge (for a character) is the minimum cost of the cartesian product of the resolutions of the adjacent nodes.

If the edge *is* a network edge:

The re-rooting candidate cost for that edge (for a character) is the minimum cost of the cartesian product of the resolutions of the adjacent nodes minus any resolutions that contain the incident network edge contained on the current network edge.

assignPunitiveNetworkEdgeCost :: (HasBlockCost u v w x y z i r, HasRootCost u v w x y z r, HasTraversalFoci z (Maybe TraversalFoci)) => PhylogeneticDAG2 e n u v w x y z -> (NonEmpty (TraversalTopology, r, r, r, Vector (NonEmpty TraversalFocusEdge)), PhylogeneticDAG2 e n u v w x y z)

Calculate and assign the punitive network edge cost for the DAG.

Consider a network \(N = (V, E)\), as commonly defined with an edge set \(E\) and vertex set \(V\). Let this network be a directed, acyclic graph (DAG) with a single root. Furthermore, consider the set of display trees \(T\), with individual display trees denoted as \(τ \in T\), derived from the resolutions of network edges in \(E\) with \(n\) leaf taxa. For a set of \(k\) character blocks \(C = \(C_1, ... , C_k\)), there is at least one most parsimonious display tree \(τ_{min} \in T\) with a cost of \(cost \left(τ_{min} \right)\) and with edge set \(E_{min} \subseteq E\) and vertex set \(V_{min} \subseteq V\). We further denote the display tree with minimum cost \(c_i\) for the character block \(C_i \in C\) as \(τ_i \in T\) with the corresponding edge set \(E_i \subseteq E\) and vertex set \(V_i \subseteq V\).

The network edge punative cost is \(\infty\) if there is an "unused" network edge in the DAG. There exists an "unused" edge in the DAG if and only if there exists a network edge \(e \in E\) such that:

\[ \forall C_i \in C \quad τ_i \in T \; \text{is the minimal display tree for} \; C_i \text{and} \; e \not \in τ_i \]

If there does not exist such a network edge we consider all the network edges in the DAG "used" and the punative network edge cost is defined as follows:

\[ \frac {\Sigma_{1}^{k} \; c_i \times | E_i \setminus E_{min} |} {2 \times \left( 2n -2 \right)} \]

For more detail on the quantification of a network with a single root, see:

Phylogenetic network analysis as a parsimony optimization problem, Wheeler 2015

We must also consider the cases of multi-rooted DAGs, extending the method described above. Unlike in the single rooted DAG context where we find a minimum display tree over all character blocks, in a multi-rooted DAG we find the "minimal display forest" over all character blocks. In the multi-rooted DAG, let \( \left\{r_1, r_2, ... r_k \right\} = \mathcal{R}\) be the set of \(k\) set roots.

The minimal display forest is the set

This function performs this punative network edge cost calculation and updates the DAG metadata to reflect the cost of the network context.

generateLocalResolutions :: HasBlockCost u'' v'' w'' x'' y'' z'' Word Double => (u -> [u'] -> u'') -> (v -> [v'] -> v'') -> (w -> [w'] -> w'') -> (x -> [x'] -> x'') -> (y -> [y'] -> y'') -> (z -> [z'] -> z'') -> ResolutionInformation (CharacterSequence u v w x y z) -> [ResolutionInformation (CharacterSequence u' v' w' x' y' z')] -> ResolutionInformation (CharacterSequence u'' v'' w'' x'' y'' z'')

Given a pre-order transformation for each type parameter, apply the transformations to each possible resolution that is not inconsistent.

postorderSequence' :: HasBlockCost u' v' w' x' y' z' Word Double => (u -> [u'] -> u') -> (v -> [v'] -> v') -> (w -> [w'] -> w') -> (x -> [x'] -> x') -> (y -> [y'] -> y') -> (z -> [z'] -> z') -> PhylogeneticDAG2 e n u v w x y z -> PhylogeneticDAG2 e n u' v' w' x' y' z'

Applies a traversal logic function over a ReferenceDAG in a pre-order manner.

The logic function takes a current node decoration, a list of parent node decorations with the logic function already applied, and returns the new decoration for the current node.

preorderFromRooting :: (HasBlockCost u v w x y z Word Double, HasBlockCost u' v' w' x' y' z' Word Double, HasTraversalFoci z (Maybe TraversalFoci), HasTraversalFoci z' (Maybe TraversalFoci)) => (z -> [(Word, z')] -> z') -> HashMap EdgeReference (ResolutionCache (CharacterSequence u v w x y z)) -> Vector (HashMap EdgeReference (ResolutionCache (CharacterSequence u v w x y z))) -> PhylogeneticDAG2 e' n' u' v' w' x' y' z -> PhylogeneticDAG2 e' n' u' v' w' x' y' z'

Applies a traversal logic function over a ReferenceDAG in a pre-order manner.

The logic function takes a current node decoration, a list of parent node decorations with the logic function already applied, and returns the new decoration for the current node.

preorderFromRooting'' :: HasTraversalFoci z' (Maybe TraversalFoci) => (z -> [(Word, z')] -> z') -> HashMap EdgeReference (ResolutionCache (CharacterSequence u v w x y z)) -> Vector (HashMap EdgeReference (ResolutionCache (CharacterSequence u v w x y z))) -> NonEmpty (TraversalTopology, Double, Double, Double, Vector (NonEmpty TraversalFocusEdge)) -> PhylogeneticDAG2 e' n' u' v' w' x' y' z -> PhylogeneticDAG2 e' n' u' v' w' x' y' z'

Applies a traversal logic function over a ReferenceDAG in a pre-order manner.

The logic function takes a current node decoration, a list of parent node decorations with the logic function already applied, and returns the new decoration for the current node.

preorderSequence' :: HasBlockCost u v w x y z Word Double => (u -> [(Word, u')] -> u') -> (v -> [(Word, v')] -> v') -> (w -> [(Word, w')] -> w') -> (x -> [(Word, x')] -> x') -> (y -> [(Word, y')] -> y') -> (z -> [(Word, z')] -> z') -> PhylogeneticDAG2 e n u v w x y z -> PhylogeneticDAG2 e n u' v' w' x' y' z'

Applies a traversal logic function over a ReferenceDAG in a pre-order manner.

The logic function takes a current node decoration, a list of parent node decorations with the logic function already applied, and returns the new decoration for the current node.

preorderSequence'' :: HasBlockCost u v w x y z Word Double => (u -> [(Word, u')] -> u') -> (v -> [(Word, v')] -> v') -> (w -> [(Word, w')] -> w') -> (x -> [(Word, x')] -> x') -> (y -> [(Word, y')] -> y') -> (z -> [(Word, z')] -> z') -> PhylogeneticDAG2 e n u v w x y z -> PhylogeneticDAG2 e n u' v' w' x' y' z'

Applies a traversal logic function over a ReferenceDAG in a pre-order manner.

The logic function takes a current node decoration, a list of parent node decorations with the logic function already applied, and returns the new decoration for the current node.

The better version.*

renderSummary :: (Foldable f, HasBlockCost u v w x y z Word Double, HasCharacterName u CharacterName, HasCharacterName v CharacterName, HasCharacterName w CharacterName, HasCharacterName x CharacterName, HasCharacterName y CharacterName, HasCharacterName z CharacterName, HasTraversalFoci z (Maybe TraversalFoci)) => PhylogeneticDAG2 e (f String) u v w x y z -> String

Nicely show the DAG information.

rootCosts :: HasBlockCost u v w x y z i r => PhylogeneticDAG2 s t u v w x y z -> NonEmpty r

Calculate the root cost for each character block.

totalEdgeCosts :: (EncodableDynamicCharacter c, Exportable c, Exportable (Element c), HasCharacterWeight z r, HasDenseTransitionCostMatrix z (Maybe DenseTransitionCostMatrix), HasSingleDisambiguation z c, HasSparseTransitionCostMatrix z MemoizedCostMatrix, NFData r, Num r, Ord (Element c)) => PhylogeneticDAG2 e n u v w x y z -> NonEmpty [r]

Computes the total edge cost over all the disambiguated final assignments.

getDotContextWithBaseAndIndex

Arguments

:: Foldable f

=> Int
Base over which the Unique

-> Int

-> PhylogeneticDAG2 e (f String) u v w x y z

-> ([DotNode GraphID], [DotEdge GraphID])

Get the dot context of a PhylogeneticDAG with useful internal node decorations.

Mapping over networks

edgePreorderMap :: e -> [e'] -> e'

Map over each edge in a pre-order manner.

edgePostorderMap :: e -> [e'] -> e'

Map over each edge in a post-order manner.

edgePreorderFold :: e -> [a] -> a

Fold over each edge in a pre-order manner.

edgePostorderFold :: e -> [a] -> a

Fold over each edge in a pre-order manner.

nodePreorderMap :: n -> [n'] -> n'

Map over each node in a pre-order manner.

nodePostorderMap :: n -> [n'] -> n'

Map over each node in a post-order manner.

nodePreorderFold :: n -> [a] -> a

Fold over each node in a pre-order manner.

nodePostorderFold :: n -> [a] -> a

Fold over each node in a post-order manner.