Bio.Graph.ReferenceDAG

Description

Efficient access

data ReferenceDAG d e n

A constant time access representation of a directed acyclic graph.

data GraphData d

Annotations which are global to the graph

- Type annotations:

d = graph metadata

Constructors

GraphData

Fields

dagCost :: !ExtendedReal

networkEdgeCost :: !ExtendedReal

rootingCost :: !Double

totalBlockCost :: !Double

graphMetadata :: d

data NodeRef

A reference to a node within the ReferenceDAG.

nodeFoldMap :: Monoid m => (n -> m) -> ReferenceDAG d e n -> m

Use the supplied transformation to fold the Node values of the DAG into a Monoid result. The fold is *loosely* ordered from *a* root node towards the leaves.

nodePostOrder :: (n -> [n'] -> n') -> ReferenceDAG d e n -> ReferenceDAG d e n'

Applies a traversal logic function over a ReferenceDAG in a post-order manner.

The logic function takes a current node decoration, a list of child node decorations with the logic function already applied, and returns the new decoration for the current node.

nodePreOrder :: (n -> [(Word, n')] -> n') -> ReferenceDAG d e n -> ReferenceDAG d e n'

Applies a traversal logic function over a ReferenceDAG in a pre-order manner.

The logic function takes a current node decoration, a list of parent node decorations with the logic function already applied, and returns the new decoration for the current node.

Effciently construct a DAG

fromList :: Foldable f => f (IntSet, n, IntMap e) -> ReferenceDAG () e n

O(n)

Constructs a ReferenceDAG from a "list" of references.

Generally regarded as *UNSAFE*! Prefer unfoldDAG.

Does not check for connectivity of the DAG.

Does not check for the lack of cycles.

Does not normalize nodes for propper in-degree values.

Does not normalize nodes for propper out-degree values.

unfoldDAG

Arguments

:: (Eq a, Hashable a, Monoid e, Monoid n)

=> (a -> ([(e, a)], n, [(e, a)]))
Unfolding function

-> a
Seed value

-> ReferenceDAG () e n

Probably, hopefully O(n).

Build the graph functionally from a generating function.

The generating function produces three results:

A list of parent edge decorations and corresponding ancestral values

The node decoration for the input value

A list of child edge decorations and corresponding descendent values

Remove metadata information and replace with a default

defaultGraphMetadata :: Monoid m => GraphData d -> GraphData m

Set the metadata to a "default" value.

Default in the function's name is used as a verb, not a noun.

defaultMetadata :: Monoid m => ReferenceDAG d e n -> ReferenceDAG m e n

Overwrite the current graph metadata with a default value.

Default in the function's name is used as a verb, not a noun.

Edgeset query

candidateNetworkEdges :: ReferenceDAG d e n -> Set ((Int, Int), (Int, Int))

Generate the set of candidate network edges for a given DAG.

referenceEdgeSet :: ReferenceDAG d e n -> EdgeSet (Int, Int)

Produces a set of directed references representing all edges in the DAG. Equivelant to:

referenceTreeEdgeSet dag `union` referenceTreeEdgeSet dag

referenceTreeEdgeSet :: ReferenceDAG d e n -> EdgeSet (Int, Int)

Produces a set of directed references representing all tree edges in the DAG. Omits network edges in the DAG. The resulting EdgeSet may not be connected.

Equivelant to:

referenceEdgeSet dag `difference` referenceNetworkEdgeSet dag

The follwoing will always hold:

null (referenceTreeEdgeSet dag `intersecttion` referenceTreeEdgeSet dag)

referenceNetworkEdgeSet :: ReferenceDAG d e n -> EdgeSet (Int, Int)

Produces a set of directed references representing all Netowrk edges in the DAG. Omits tree edges in the DAG. The resulting EdgeSet *will not* be connected.

Equivelant to:

referenceEdgeSet dag `difference` referenceTreeEdgeSet dag

The follwoing will always hold:

null (referenceTreeEdgeSet dag `intersecttion` referenceTreeEdgeSet dag)

undirectedRootEdgeSet :: ReferenceDAG d e n -> EdgeSet (Int, Int)

Produces a set of undirected references representing all undirected edges that have a root edges in the DAG, if the root(s) were removed an to create an undirected network. Omits all edges in the undirected network representation of the DAG that are not specified as a root of the DAG. The resulting EdgeSet (unless of trivial cardinality) *will not* be connected.

Edgeset manipulation

connectEdge

Arguments

:: Monoid e

=> ReferenceDAG d e n

-> (n -> n -> n -> n)
Function describing how to construct the new origin internal node: parent, existing child, new child

-> (n -> n -> n)
Function describing how to construct the new target internal node: exisiting parent, exisiting child

-> (Int, Int)
Origin edge (coming from)

-> (Int, Int)
Target edge (going too)

-> ReferenceDAG d e n

Given two edges, and two functions describing the way to construct new node values, adds a new edge between the two supplied edges and constructs two intermediary nodes.

invadeEdge

Arguments

:: Monoid e

=> ReferenceDAG d e n

-> (n -> n -> n -> n)
Function describing how to construct the new internal node, parent, old child, new child

-> n

-> (Int, Int)

-> ReferenceDAG d e n

Add a node on the supplied edge, creating two new incident edges.

Rendering

referenceRendering :: ReferenceDAG d e n -> String

Renders the ReferenceDAG without showing the node or edge decorations. Displays a multi-line, tabular reference map of the ReferenceDAG.

topologyRendering :: ReferenceDAG d e n -> String

Displays a tree-like rendering of the ReferenceDAG.

getDotContext

Arguments

:: Foldable f

=> Int
Base over which the Unique

-> Int

-> ReferenceDAG d e (f String)

-> ([DotNode GraphID], [DotEdge GraphID])

Exctract a context from the ReferenceDAG that can be used to create a dot context for rendering.