Bio.Graph.ZipperDAG

Description

Mutually recursive zipper types

data ZipperNode e n

A node in the zipper structure.

The node's decoration value can be accessed through the Cursor instance functions.

The node's edges can bew extracted with zipperEdges.

data ZipperEdge n e

An edge of the zipper structure.

The edge's decoration value can be accessed through the Cursor instance functions.

The edge's parent node can be accessed with zipperEdgeParent.

The edge's child node can be accessed with zipperEdgeChild.

Zipper cursor accessors

class Cursor f where

Zipper types with an accessible cursor. The instance methods should be implemented in constant time for this type-class to be useful.

Minimal complete definition

getCursor, (setCursor | adjustCursor)

Methods

adjustCursor :: (a -> a) -> f a -> f a

Monomorphically modify the value at the zipper's current cursor position with the supplied function.

getCursor :: f a -> a

Retreive the value at the zipper's current cursor position.

setCursor :: a -> f a -> f a

Overwrite the value at the zipper's current cursor position.

Zipper neighbor accessors

zipperEdgeChild :: ZipperEdge n e -> ZipperNode e n

O(1) Retrieves the child ZipperNode of a ZipperEdge

zipperEdgeParent :: ZipperEdge n e -> ZipperNode e n

O(1)

Retrieves the parent ZipperNode of a ZipperEdge

zipperEdges :: ZipperNode e n -> [ZipperEdge n e]

O(e)

Retrieves each ZipperEdge of a ZipperNode.

Construction

unfoldDAG :: (b -> ([(e, b)], n, [(e, b)])) -> b -> ZipperNode e n

Build the graph functionally from a generating function.