Bio.Metadata.CharacterName

Description

A CharacterName represents an unique name identifier for a character that can be used as a textual reference.

A CharacterName contains within it a reference to the file from which it was originally parsed, allowing identically named characters from different input files to be uniquely referenced. The file from which a CharacterName was originally parsed can be queried by the function sourceFile.

A CharcterName can be a either a user specified textual identifier or a default generated value. Where a given CharacterName is user specified or defaulted may be queried by the function isUserDefined.

A CharacterName is only constructable by calling makeCharacterNames. Calls to makeCharacterNames should supply all possible character information available at parse time to ensure that the resulting CharacterNames are uniquely idenitfible.

A CharacterName has an ordering defined as follows:

The lexical ordering of the CharacterName source file name

User defined values take precedence over a defaulted values within the same source file

The lexical ordering of the user defined textual identifier within the same source file

The numeric ordering of the defaulted values within the same source file.

Documentation

data CharacterName

Represents the name of a character in a type-safe manner which resolves namespace ambiguities.

isUserDefined :: CharacterName -> Bool

Determine if the CharacterName was user defined or defaulted.

makeCharacterNames :: Traversable t => t (FilePath, Maybe String) -> t CharacterName

Construct many CharacterNames for the input list of '(FilePath, Maybe String)' pairs. Just values represent user supplied character names while Nothing values represent character names to be defaulted. However, in some cases Just valued strings may be rejected and defaulted instead. This occurs when the user defined character name starts with a ":" or is empty.

The construction is order preserving on the input structure. Multiple characters from a ffrom the same input file will be defualted with incrementing indices.

Examples

Basic usage:

>>> makeCharacterNames [("file.dat", Nothing)]
["file.dat:0"]

>>> makeCharacterNames [("file.dat", Just "Eyes")]
["Eyes"]

>>> makeCharacterNames [("path/to/file.dat", Nothing")]
["file.dat:0"]

>>> makeCharacterNames [("path/to/file.dat", Just "Eyes")]
["Eyes"]

>>> makeCharacterNames [("file.dat", Nothing), ("file.dat", Nothing)]
["file.dat:0","file.dat:1"]

>>> makeCharacterNames [("file.dat", Nothing), ("file.dat", Just "Eyes"), ("file.dat", Nothing)]
["file.dat:0","Eyes","file.dat:1"]

>>> makeCharacterNames [("foo.txt", Nothing), ("foo.txt", Nothing), ("bar.txt", Nothing), ("baz.txt", Nothing), ("baz.txt", Nothing)]
["foo.txt:0","foo.txt:1","bar.txt:0","baz.txt:0","baz.txt"]

>>> makeDefaultCharacterNameRange ("virus.exe", Just "")
["virus.exe:0"]

>>> makeDefaultCharacterNameRange ("virus.exe", Just ":")
["virus.exe:0"]

>>> makeDefaultCharacterNameRange ("virus.exe", Just ":a")
["virus.exe:0"]

>>> makeDefaultCharacterNameRange ("virus.exe", Just "a:")
["a:"]

>>> makeDefaultCharacterNameRange ("virus.exe", Just "%#:!")
["%#:!"]

>>> makeCharacterNames [("foo.txt", Nothing), ("foo.tx", Just ""), ("foo.tx", Nothing)]
["foo.txt:0","foo.txt:1","foo.txt:2"]

sourceFile :: CharacterName -> FilePath

Determine the source file for the character.