Bio.Metadata.Discrete

Description

Documentation

data DiscreteCharacterMetadataDec

Represents a concrete type containing metadata fields shared across all discrete different bins. Continous bins do not have Alphabets.

class (GeneralCharacterMetadata s, HasCharacterAlphabet s (Alphabet String)) => DiscreteCharacterMetadata s where

A decoration of an initial encoding of a dynamic character which has the appropriate Lens & character class constraints.

Minimal complete definition

extractDiscreteCharacterMetadata

Methods

extractDiscreteCharacterMetadata :: s -> DiscreteCharacterMetadataDec

class (HasCharacterName s CharacterName, HasCharacterWeight s Double) => GeneralCharacterMetadata s where

A constraint for type containing metadata.

Minimal complete definition

extractGeneralCharacterMetadata

Methods

extractGeneralCharacterMetadata :: s -> GeneralCharacterMetadataDec

class HasCharacterAlphabet s a | s -> a where

A Lens for the alphabet field

Minimal complete definition

characterAlphabet

Methods

characterAlphabet :: Lens' s a

class HasCharacterName s a | s -> a where

A Lens for the characterName field

Minimal complete definition

characterName

Methods

characterName :: Lens' s a

class HasCharacterWeight s a | s -> a where

A Lens for the characterWeight field

Minimal complete definition

characterWeight

Methods

characterWeight :: Lens' s a

discreteMetadata :: CharacterName -> Double -> Alphabet String -> DiscreteCharacterMetadataDec

Construct a concrete typed DiscreteCharacterMetadataDec value from the supplied inputs.