Bio.Metadata.DiscreteWithTCM

Description

Documentation

class (GeneralCharacterMetadata s, HasCharacterAlphabet s (Alphabet String)) => DiscreteCharacterMetadata s where

A decoration of an initial encoding of a dynamic character which has the appropriate Lens & character class constraints.

Minimal complete definition

extractDiscreteCharacterMetadata

Methods

extractDiscreteCharacterMetadata :: s -> DiscreteCharacterMetadataDec

class (DiscreteCharacterMetadata s, EncodableStreamElement c, HasSymbolChangeMatrix s (Word -> Word -> Word), HasTransitionCostMatrix s (c -> c -> (c, Word))) => DiscreteWithTcmCharacterMetadata s c | s -> c

A decoration of an initial encoding of a dynamic character which has the appropriate Lens & character class constraints.

data DiscreteWithTCMCharacterMetadataDec c

Represents a concrete type containing metadata fields shared across all discrete different bins. Continous bins do not have Alphabets.

class (HasCharacterName s CharacterName, HasCharacterWeight s Double) => GeneralCharacterMetadata s where

A constraint for type containing metadata.

Minimal complete definition

extractGeneralCharacterMetadata

Methods

extractGeneralCharacterMetadata :: s -> GeneralCharacterMetadataDec

class HasCharacterAlphabet s a | s -> a where

A Lens for the alphabet field

Minimal complete definition

characterAlphabet

Methods

characterAlphabet :: Lens' s a

class HasCharacterName s a | s -> a where

A Lens for the characterName field

Minimal complete definition

characterName

Methods

characterName :: Lens' s a

class HasCharacterWeight s a | s -> a where

A Lens for the characterWeight field

Minimal complete definition

characterWeight

Methods

characterWeight :: Lens' s a

class HasSymbolChangeMatrix s a | s -> a where

A Lens for the symbolChangeMatrix field

Minimal complete definition

symbolChangeMatrix

Methods

symbolChangeMatrix :: Lens' s a

class HasTransitionCostMatrix s a | s -> a where

A Lens for the transitionCostMatrix field

Minimal complete definition

transitionCostMatrix

Methods

transitionCostMatrix :: Lens' s a

class HasSparseTransitionCostMatrix s a | s -> a where

A Lens for the denseTransitionCostMatrix field

Minimal complete definition

sparseTransitionCostMatrix

Methods

sparseTransitionCostMatrix :: Lens' s a

discreteMetadataFromTCM :: CharacterName -> Double -> Alphabet String -> TCM -> DiscreteWithTCMCharacterMetadataDec c

Construct a concrete typed DiscreteWithTCMCharacterMetadataDec value from the supplied inputs.

discreteMetadataWithTCM :: CharacterName -> Double -> Alphabet String -> (Word -> Word -> Word) -> DiscreteWithTCMCharacterMetadataDec c

Construct a concrete typed DiscreteWithTCMCharacterMetadataDec value from the supplied inputs.