Bio.Metadata.Dynamic

Description

Documentation

data DenseTransitionCostMatrix

Exposed wrapper for C allocated cost matrix structs.

class (DiscreteWithTcmCharacterMetadata s c, HasDenseTransitionCostMatrix s (Maybe DenseTransitionCostMatrix), HasSparseTransitionCostMatrix s MemoizedCostMatrix, HasTraversalFoci s (Maybe TraversalFoci)) => DynamicCharacterMetadata s c | s -> c where

A decoration of an initial encoding of a dynamic character which has the appropriate Lens & character class constraints.

Minimal complete definition

extractDynamicCharacterMetadata

Methods

extractDynamicCharacterMetadata :: s -> DynamicCharacterMetadataDec c

data DynamicCharacterMetadataDec c

Represents a concrete type containing metadata fields shared across all discrete different bins. Continous bins do not have Alphabets.

class HasCharacterAlphabet s a | s -> a where

A Lens for the alphabet field

Minimal complete definition

characterAlphabet

Methods

characterAlphabet :: Lens' s a

class HasCharacterName s a | s -> a where

A Lens for the characterName field

Minimal complete definition

characterName

Methods

characterName :: Lens' s a

class HasCharacterWeight s a | s -> a where

A Lens for the characterWeight field

Minimal complete definition

characterWeight

Methods

characterWeight :: Lens' s a

class HasDenseTransitionCostMatrix s a | s -> a where

A Lens for the denseTransitionCostMatrix field

Minimal complete definition

denseTransitionCostMatrix

Methods

denseTransitionCostMatrix :: Lens' s a

class HasSymbolChangeMatrix s a | s -> a where

A Lens for the symbolChangeMatrix field

Minimal complete definition

symbolChangeMatrix

Methods

symbolChangeMatrix :: Lens' s a

class HasTransitionCostMatrix s a | s -> a where

A Lens for the transitionCostMatrix field

Minimal complete definition

transitionCostMatrix

Methods

transitionCostMatrix :: Lens' s a

class HasTraversalFoci s a | s -> a where

A Lens for the traversalFoci field

Minimal complete definition

traversalFoci

Methods

traversalFoci :: Lens' s a

data MemoizedCostMatrix

A type-safe wrapper for the mutable, memoized TCm.

type TraversalFoci = NonEmpty TraversalFocus

Represents a collection of paired rooting edges and unrooted topologies.

type TraversalFocusEdge = (Int, Int)

An unrooted edge.

type TraversalTopology = TopologyRepresentation TraversalFocusEdge

A unique representation of a DAG topology.

dynamicMetadata :: CharacterName -> Double -> Alphabet String -> (Word -> Word -> Word) -> Maybe DenseTransitionCostMatrix -> DynamicCharacterMetadataDec c

Construct a concrete typed DynamicCharacterMetadataDec value from the supplied inputs.

dynamicMetadataFromTCM :: CharacterName -> Double -> Alphabet String -> TCM -> DynamicCharacterMetadataDec c

Construct a concrete typed DynamicCharacterMetadataDec value from the supplied inputs.

maybeConstructDenseTransitionCostMatrix :: Alphabet a -> (Word -> Word -> Word) -> Maybe DenseTransitionCostMatrix

O(n^3) Constructs 2D & 3D dense TCMs.

Conditionally construct a DenseTransitionCostMatrix. If the alphabet size is too large, a Nothing value will be returned. Otherwise the dense TCM is constructed strictly at the time the function in invoked.

Currentlty returns a Just value for alphabet sizes in the range [2..8].