Bio.Sequence

Description

Data structures and instances for coded characters Coded characters are dynamic characters recoded as

Documentation

data CharacterSequence u v w x y z

A multi-level partitioned, non-empty sequence of characters.

A sequence is partitioned into blocks, each block is partitioned into static and dynamic charcters, with the static characters partitioned by optimization routines.

Blocks are optimized atomically with resepect to network resolutions.

Some blocks may not exist on a given node after performing a pre-order traversal. This occurs when the minimal display tree for a given block does not include all the nodes in the DAG. In this case the nodes not incldued in the topology will have "missing" final assignments for all blocks for which that topology was minimal.

data CharacterBlock u v w x y z

Represents a block of charcters which are optimized atomically together across networks. The CharacterBlock is polymorphic over static and dynamic character definitions.

Bin types: * u = * v = * w = * x = * y = * z =

Use '(<>)' to construct larger blocks.

Constructors

CharacterBlock

Fields

continuousCharacterBins :: Vector u

nonAdditiveCharacterBins :: Vector v

additiveCharacterBins :: Vector w

metricCharacterBins :: Vector x

nonMetricCharacterBins :: Vector y

dynamicCharacters :: Vector z

type HasBlockCost u v w x y z i r = (HasCharacterCost u r, HasCharacterCost v i, HasCharacterCost w i, HasCharacterCost x i, HasCharacterCost y i, HasCharacterCost z i, HasCharacterWeight u r, HasCharacterWeight v r, HasCharacterWeight w r, HasCharacterWeight x r, HasCharacterWeight y r, HasCharacterWeight z r, Integral i, Real r)

CharacterBlocks satisfying this constraint have a calculable cost.

type HasRootCost u v w x y z r = (HasCharacterWeight u r, HasCharacterWeight v r, HasCharacterWeight w r, HasCharacterWeight x r, HasCharacterWeight y r, HasCharacterWeight z r, HasAverageLength z AverageLength, PossiblyMissingCharacter u, PossiblyMissingCharacter v, PossiblyMissingCharacter w, PossiblyMissingCharacter x, PossiblyMissingCharacter y, PossiblyMissingCharacter z, Floating r, Real r)

CharacterBlocks satisfying this constraint have a calculable cost.

data PartialCharacterBlock u v w x y z

Represents a block of charcters which are optimized atomically together across networks. The CharacterBlock is polymorphic over static and dynamic character definitions.

Use '(<>)' to construct larger blocks.

continuousSingleton :: c -> PartialCharacterBlock c v w x y z

Construct a singleton block containing a continuous character.

discreteSingleton :: TCMStructure -> s -> PartialCharacterBlock c s s s s d

Construct a singleton block containing a discrete character.

dynamicSingleton :: d -> PartialCharacterBlock u v w x y d

Construct a singleton block containing a dynamic character.

finalizeCharacterBlock :: PartialCharacterBlock u v w x y z -> CharacterBlock u v w x y z

Converts a PartialCharacterBlock to a CharacterBlock, finalizing the efficient construction process.

toMissingCharacters :: (PossiblyMissingCharacter u, PossiblyMissingCharacter v, PossiblyMissingCharacter w, PossiblyMissingCharacter x, PossiblyMissingCharacter y, PossiblyMissingCharacter z) => CharacterBlock u v w x y z -> CharacterBlock u v w x y z

Convert all characters contained in the block to thier missing value.

toBlocks :: CharacterSequence u v w x y z -> NonEmpty (CharacterBlock u v w x y z)

Destructs a CharacterSequence to it's composite blocks.

fromBlocks :: NonEmpty (CharacterBlock u v w x y z) -> CharacterSequence u v w x y z

Constructs a CharacterSequence from a non-empty colection of blocks.

toBlockVector :: CharacterSequence u v w x y z -> Vector (CharacterBlock u v w x y z)

Destructs a CharacterSequence to it's composite blocks.

fromBlockVector :: Vector (CharacterBlock u v w x y z) -> CharacterSequence u v w x y z

Destructs a CharacterSequence to it's composite blocks.

hexmap :: (u -> u') -> (v -> v') -> (w -> w') -> (x -> x') -> (y -> y') -> (z -> z') -> CharacterSequence u v w x y z -> CharacterSequence u' v' w' x' y' z'

Perform a six way map over the polymorphic types.

hexTranspose :: Traversable1 t => t (CharacterSequence u v w x y z) -> CharacterSequence [u] [v] [w] [x] [y] [z]

Performs a 2D transform on the Traversable structure of CharacterSequence values.

Assumes that the CharacterSequence values in the Traversable structure are of equal length. If this assumtion is violated, the result will be truncated.

hexZipWith :: (u -> u' -> u'') -> (v -> v' -> v'') -> (w -> w' -> w'') -> (x -> x' -> x'') -> (y -> y' -> y'') -> (z -> z' -> z'') -> CharacterSequence u v w x y z -> CharacterSequence u' v' w' x' y' z' -> CharacterSequence u'' v'' w'' x'' y'' z''

Performs a zip over the two character sequences. Uses the input functions to zip the different character types in the character block.

Assumes that the CharacterSequence values have the same number of character blocks and the same number of each character type in the corresponding block of each block. If this assumtion is violated, the result will be truncated.

sequenceCost :: HasBlockCost u v w x y z i r => CharacterSequence u v w x y z -> r

Calculates the cumulative cost of a CharacterSequence. Performs some of the operation in parallel.

sequenceRootCost :: (HasRootCost u v w x y z r, Integral i) => i -> CharacterSequence u v w x y z -> r

Calculates the root cost of a CharacterSequence. Performs some of the operation in parallel.

blockCost :: HasBlockCost u v w x y z i r => CharacterBlock u v w x y z -> r

Calculates the cost of a CharacterBlock. Performs some of the operation in parallel.

staticCost :: HasBlockCost u v w x y z i r => CharacterBlock u v w x y z -> r

Calculates the cost of a CharacterBlock. Performs some of the operation in parallel.