Bio.Sequence.Block

Description

Documentation

data CharacterBlock u v w x y z

Represents a block of charcters which are optimized atomically together across networks. The CharacterBlock is polymorphic over static and dynamic character definitions.

Bin types: * u = * v = * w = * x = * y = * z =

Use '(<>)' to construct larger blocks.

Constructors

CharacterBlock

Fields

continuousCharacterBins :: Vector u

nonAdditiveCharacterBins :: Vector v

additiveCharacterBins :: Vector w

metricCharacterBins :: Vector x

nonMetricCharacterBins :: Vector y

dynamicCharacters :: Vector z

type HasBlockCost u v w x y z i r = (HasCharacterCost u r, HasCharacterCost v i, HasCharacterCost w i, HasCharacterCost x i, HasCharacterCost y i, HasCharacterCost z i, HasCharacterWeight u r, HasCharacterWeight v r, HasCharacterWeight w r, HasCharacterWeight x r, HasCharacterWeight y r, HasCharacterWeight z r, Integral i, Real r)

CharacterBlocks satisfying this constraint have a calculable cost.

type HasRootCost u v w x y z r = (HasCharacterWeight u r, HasCharacterWeight v r, HasCharacterWeight w r, HasCharacterWeight x r, HasCharacterWeight y r, HasCharacterWeight z r, HasAverageLength z AverageLength, PossiblyMissingCharacter u, PossiblyMissingCharacter v, PossiblyMissingCharacter w, PossiblyMissingCharacter x, PossiblyMissingCharacter y, PossiblyMissingCharacter z, Floating r, Real r)

CharacterBlocks satisfying this constraint have a calculable cost.

blockCost :: HasBlockCost u v w x y z i r => CharacterBlock u v w x y z -> r

Calculates the cost of a CharacterBlock. Performs some of the operation in parallel.

rootCost :: (HasRootCost u v w x y z r, Integral i) => i -> CharacterBlock u v w x y z -> r

Calculate the "rooting cost" of a CharacterBlock by applying a "rooting- multiplier" based on the number of other roots in the DAG.

staticCost :: HasBlockCost u v w x y z i r => CharacterBlock u v w x y z -> r

Calculates the cost of a CharacterBlock. Performs some of the operation in parallel.

toMissingCharacters :: (PossiblyMissingCharacter u, PossiblyMissingCharacter v, PossiblyMissingCharacter w, PossiblyMissingCharacter x, PossiblyMissingCharacter y, PossiblyMissingCharacter z) => CharacterBlock u v w x y z -> CharacterBlock u v w x y z

Convert all characters contained in the block to thier missing value.

hexmap :: (u -> u') -> (v -> v') -> (w -> w') -> (x -> x') -> (y -> y') -> (z -> z') -> CharacterBlock u v w x y z -> CharacterBlock u' v' w' x' y' z'

Perform a six way map over the polymorphic types.

hexTranspose :: Traversable t => t (CharacterBlock u v w x y z) -> CharacterBlock (t u) (t v) (t w) (t x) (t y) (t z)

Performs a 2D transform on the Traversable structure of CharacterBlock values.

Assumes that the CharacterBlock values in the Traversable structure are of equal length. If this assumtion is violated, the result will be truncated.

hexZipWith :: (u -> u' -> u'') -> (v -> v' -> v'') -> (w -> w' -> w'') -> (x -> x' -> x'') -> (y -> y' -> y'') -> (z -> z' -> z'') -> CharacterBlock u v w x y z -> CharacterBlock u' v' w' x' y' z' -> CharacterBlock u'' v'' w'' x'' y'' z''

Performs a zip over the two character blocks. Uses the input functions to zip the different character types in the character block.

Assumes that the CharacterBlock values have the same number of each character type. If this assumtion is violated, the result will be truncated.