Control.Evaluation

Description

The Evaluation type for representing a computational context.

Documentation

data EvaluationT m a

A monad transformer of Evaluation.

data Evaluation a

Reprsents a monoidal evaluations of a value a with contextual notifications. Stores a list of ordered contextual notifications retrievable by notifications. Holds the evaluation state a retrievable evaluationResult.

data EvalUnit a

The internal state of the computation. A short-circuiting evaluation unit which returns a value, and error, or indicated that no work was done

Constructors

NoOp

Error String

Value a

data Notification

Represents the types of contextual notifications that can be generated by the evaluation.

evalEither :: Show s => Either s b -> Evaluation b

Lifts an Either with a Show error condition into the Evaluation context.

evalIO :: IO a -> EvaluationT IO a

Synonym for impure

evaluationResult :: Evaluation a -> EvalUnit a

Retrieve the result state from the Evaluation.

impure :: IO a -> EvaluationT IO a

Takes an IO value and lifts it into the evaluation context.

notifications :: Evaluation a -> [Notification]

Retrieve the ordered list of contextual Notifications from the Evaluation.

runEvaluation :: EvaluationT m a -> m (Evaluation a)

Run the EvaluationT monad transformer

state :: Monad m => Evaluation a -> EvaluationT m a

Takes an Evaluation and lifts it into the transformer context.

showRun :: Show a => EvaluationT IO a -> IO ()

Prints an IO parameterized transformer of Evaluation context to the STDOUT.