Control.Parallel.Custom

Description

Documentation

parmap :: Traversable t => Strategy b -> (a -> b) -> t a -> t b

Map a function over a traversable structure in parrallel. Should be prefered over parMap which is fixed to lists.

parZipWith :: (Traversable t, Zip t) => Strategy c -> (a -> b -> c) -> t a -> t b -> t c

Zip two traversable, zippable structures in parrallel with a function.