Data.Alphabet

Description

An Alphabet represents an ordered list of unique symbols with constant time random access. Symbols are any data type which are coercable from a String through the IsString type-class.

An Alphabet is constructed in one of two ways:

Supplying a Foldable structure of symbols which are IsString instances to the fromSymbols function.

Supplying a Foldable structure of symbols and state name pairs, both of which are IsString instances to the fromSymbolsWithStateNames function.

Both Alphabet construction methods are order preserving with respect to the input symbol order.

Every Alphabet contains a "gap" symbol denoted by the expression: > fromString "-" The "gap" character is always the last element in the ordered list regardless of its presence or position in the construction structure.

An Alphabet will never contain the "missing" symbol denoted by the expression: > fromString "?" This symbol will be removed from the Alphabet if it is present in the construction structure.

Documentation

data Alphabet a

A collection of symbols and optional corresponding state names.

type AmbiguityGroup a = NonEmpty a

A non empty collection of symbols from an Alphabet.

Construction

fromSymbols :: (Ord a, IsString a, Foldable t) => t a -> Alphabet a

\( \mathcal{O} \left( n * \log_2 n \right) \)

Constructs an Alphabet from a Foldable structure of symbols which are IsString values.

fromSymbolsWithStateNames :: (Ord a, IsString a, Foldable t) => t (a, a) -> Alphabet a

\( \mathcal{O} \left( n * \log_2 n \right) \)

Constructs an Alphabet from a Foldable structure of symbols and coresponding state names, both of which a are IsString values.

The input ordering is preserved.

Querries

alphabetStateNames :: Alphabet a -> [a]

\( \mathcal{O} \left( n \right) \)

Retreives the state names for the symbols of the Alphabet.

If there the symbols of the Alphabet were not given state names during construction then an empty list is returned.

alphabetSymbols :: Alphabet a -> [a]

\( \mathcal{O} \left( n \right) \)

Retreives the symbols of the Alphabet. Synonym for toList.

gapSymbol :: Alphabet a -> a

\( \mathcal{O} \left( 1 \right) \)

Retreives the "gap character" from the alphabet.

Truncation

truncateAtSymbol :: (Ord a, IsString a) => a -> Alphabet a -> Alphabet a

\( \mathcal{O} \left( n * \log_2 n \right) \)

Attempts to find the symbol in the Alphabet. If the symbol exists, returns an alphabet with all the symbols occuring before the supplied symbol included and all symbols occuring after the supplied symbol excluded. The gap character is preserved in the alphabet regardless of the supplied symbol.

truncateAtMaxSymbol :: (Foldable t, Ord a, IsString a) => t a -> Alphabet a -> Alphabet a

\( \mathcal{O} \left( n * \log_2 n \right) \)

Attempts to find the maximum provided symbol in the Alphabet. If the any of the provided symbols exists, returns an alphabet including all the symbols occuring before the maximum provided symbol and all symbols occuring after the maximum supplied symbol excluded. The gap character is preserved in the alphabet regardless of the supplied symbol.