Data.Alphabet.IUPAC

Description

Documentation

isAlphabetAminoAcid :: (IsString s, Ord s) => Alphabet s -> Bool

O(n)

Determines if the supplied alphabet represents amino acid symbols.

Useful for determining if an AmbiguityGroup should be rendered as an IUPAC code.

isAlphabetDna :: (IsString s, Ord s) => Alphabet s -> Bool

O(n)

Determines if the supplied alphabet represents DNA symbols.

Useful for determining if an AmbiguityGroup should be rendered as an IUPAC code.

isAlphabetRna :: (IsString s, Ord s) => Alphabet s -> Bool

O(n)

Determines if the supplied alphabet represents DNA symbols.

Useful for determining if an AmbiguityGroup should be rendered as an IUPAC code.

iupacToAminoAcid :: Bimap (AmbiguityGroup String) (AmbiguityGroup String)

Substitutions for converting to an Amino Acid sequence based on IUPAC codes.

iupacToDna :: Bimap (AmbiguityGroup String) (AmbiguityGroup String)

Substitutions for converting to a DNA sequence based on IUPAC codes.

iupacToRna :: Bimap (AmbiguityGroup String) (AmbiguityGroup String)

Substitutions for converting to a RNA sequence based on IUPAC codes.

module Data.Alphabet