Data.EdgeLength

Description

A possible present, real-valued edge length with a Monoid instance defined in terms of addition.

Documentation

data EdgeLength

A monoidal edge annotation for representing an edge length.

The identity element represents an *unspecified* edge length, which is distinct from a specified edge length of length zero.

The assoiative semigroup opperation is isomorphic to addition over the real numbers when both operands are specified edge length. When one or more of the edge lengths are unspecified (the identity element), the semigroup operation returns a specified edge length if possible, satifying the monoid laws regarding the identity element.

fromDouble :: Double -> EdgeLength

Construct an EdgeLength from a Double value.

fromDoubleMay :: Maybe Double -> EdgeLength

Construct an EdgeLength from a 'Maybe Double' value. A Nothing value indicates aa missing edge length.