Data.TCM

Description

A matrix of transition costs between alphabet symbols. Exposes row-major monomorphic maps, folds, and traversals.

Documentation

data TCM

A data structure for storing a two dimensional, square array of dimensionality greater that or equal to two with positive cost values at the array indices. Values are stored in an unboxed structure for cache efficiency.

A TCM can be constructed by calling one of the following functions:

fromList

fromCols

fromRows

generate

Attempts to construct an empty or singleton TCM through the above constructors will result in a runtime exception.

data TCMDiagnosis

The result of a call to diagnoseTcm.

Constructors

TCMDiagnosis

Fields

factoredWeight :: Int

The multiplicative constant factor of a TCM. Minimum value of the multiplicative identiy one.

factoredTcm :: TCM

The new TCM with each value divided by the factoredWeight.

tcmStructure :: TCMStructure

The most restrictive present in the factoredTcm.

data TCMStructure

There is a heirachichal nature to a TCM's possible structure:

      NonSymmetric

          |

       Symmetric

          |

        Metric

        /    \

UltraMetric  Additive

     |

NonAdditive

Each TCM structure has certain properties that must hold and allows for space & time optimizations.

Non-Symmetric: No structure expoitable for optimizations.

Symmetric: Allows for half the space allocation.

σ(i,j) = σ(j,i)

Metric: Allows for half the space allocation & runtime optimizations

σ(i,j) = 0 iff i = j

σ(i,j) = σ(j,i)

σ(i,k) ≤ σ(i,j) + σ(j,k)

Ultra-Metric: Allows for half the space allocation & runtime optimizations

σ(i,j) = 0 iff i = j

σ(i,j) = σ(j,i)

σ(i,k) ≤ max { σ(i,j), σ(j,k) }

Non-Additive: Allows for no space allocation & runtime optimizations

σ(i,j) = 0 iff i = j

σ(i,j) = 1 iff i ≠ j

Additive: Allows for no space allocation & runtime optimizations

σ(i,j) = max(i,j) - min(i,j)

Constructors

NonSymmetric

Symmetric

Metric

UltraMetric

Additive

NonAdditive

Construction

fromCols :: (Foldable t, Foldable t', Real a) => t (t' a) -> (Rational, TCM)

O(n*n)

Construct a TCM from a list of columns.

Examples

>>> fromCols [[1,2,3],[4,5,6],[7,8,9]]
TCM: 3 x 3
  1 4 7
  2 5 8
  3 6 9

fromList :: (Foldable t, Real a) => t a -> (Rational, TCM)

O(n*n)

Construct a TCM from a list of elements in row major order.

Examples

>>> fromList [1..9]
TCM: 3 x 3
  1 2 3
  4 5 6
  7 8 9

>>> fromList []
*** Exception: fromList: An empty structure was supplied. Cannot construct an empty TCM!

>>> fromList [42]
*** Exception: fromList: A singleton structure was supplied. Cannot construct a TCM with dimension of 1, must have dimension of 2 or greater.

>>> fromList [1..12]
*** Exception: fromList: The number of element (12) is not a square number. Cannot construct an non-square TCM! The number of elements (12) lies between the valid square numbers (9) and (16).

fromRows :: (Foldable t, Foldable t', Real a) => t (t' a) -> (Rational, TCM)

O(n*n)

Construct a TCM from a list of rows.

Examples

>>> fromRows [[1,2,3],[4,5,6],[7,8,9]]
TCM: 3 x 3
  1 2 3
  4 5 6
  7 8 9

generate

Arguments

:: (Enum i, Integral a)

=> Int
Number of rows & columns in the TCM.

-> ((i, i) -> a)
Function to determine the value of a given index.

-> TCM

O(n*n)

A generating function for a TCM. Efficiently constructs a TCM of the specified dimensions with each value defined by the result of the supplied function.

Examples

>>> generate 5 $ const 5
TCM: 5 x 5
  5 5 5 5 5
  5 5 5 5 5
  5 5 5 5 5
  5 5 5 5 5
  5 5 5 5 5

>>> generate 4 $ \(i,j) -> abs (i - j)
TCM: 4 x 4
  0 1 2 3 4
  1 0 1 2 3
  2 1 0 1 2
  3 2 1 0 1
  4 3 2 1 0

>>> generate 8 $ \(i,j) -> if i == j || i + j == 6 then 0 else 1
TCM: 8 x 8
  0 1 1 1 0 1 1 1
  1 0 1 0 1 1 1 1
  1 1 0 1 1 1 1 1
  1 0 1 0 1 1 1 1
  0 1 1 1 0 1 1 1
  1 1 1 1 1 0 1 1
  1 1 1 1 1 1 0 1
  1 1 1 1 1 1 1 0

Indexing

(!) :: Enum i => TCM -> (i, i) -> Word32

O(1)

Indexing without bounds checking.

(!?) :: Enum i => TCM -> (i, i) -> Maybe Word32

O(1)

Safe indexing.

Queries

size :: TCM -> Int

O(1)

The number of rows and columns in the TCM.

Specialization Utility

diagnoseTcm :: TCM -> TCMDiagnosis

Performs a diagnosis of a TCM factoring out any multiplicative factor as a constant weighting and reporting the most restricitive structure present in the TCM.

reduceTcm :: IntSet -> TCM -> TCM

O(n^2)

Removes the supplied columns/rows from the TCM.