Data.TCM.Memoized

Description

Documentation

data MemoizedCostMatrix

A type-safe wrapper for the mutable, memoized TCm.

generateMemoizedTransitionCostMatrix

Arguments

:: Word
Alphabet size

-> (Word -> Word -> Word)
Generating function

-> MemoizedCostMatrix

O(n^2) where n is the alphabet size.

Generate a memoized TCM by supplying the size of the symbol alphabet and the generating function for unambiguous symbol change cost to produce a memoized TCM. A memoized TCM computes all the costs and medians of unambiguous, singleton symbol set transitions strictly when this function is invoked. A memoized TCM calculates the cost and medians of ambiguous symbol sets in a lazy, memoized manner.

Note:* The collection of ambiguous symbols set transitions is the powerset of the collection of unambiguous, singleton symbol sets. Laziness and memoization is requisite for efficient computation on any non-trivial alphabet size.

getMedianAndCost :: Exportable s => MemoizedCostMatrix -> s -> s -> (s, Word)

O(1) amortized.

Calculate the median symbol set and transition cost between the two input symbol sets.

Note:* This operation is lazily evaluated and memoized for future calls.