Data.TopologyRepresentation

Description

Set-like structures for collection of edges.

Documentation

data TopologyRepresentation a

Represents a collection of network edges and their mutually exclusive edges.

Often used to represent a unique spanning tree in a phylogenetic DAG.

Construction

isolatedNetworkEdgeContext

Arguments

:: Eq a

=> a
Included network edge

-> a
Excluded incident network edge

-> TopologyRepresentation a

\( \mathcal{O} \left( 1 \right) \)

Construct a singleton TopologyRepresentation value by supplying a network edge identifier representing an edge contained in the topology and a corresponding network edge identifier representing the mutually exclusive incident edge.

Use the semigroup operator '(<>)' to merge isolated network edge contexts into a larger TopologyRepresentation.

Deconstruct

excludedNetworkEdges :: TopologyRepresentation a -> Set a

\( \mathcal{O} \left( n \right) \)

Retreive the list of network edge identifiers excluded from the topology.

includedNetworkEdges :: TopologyRepresentation a -> Set a

\( \mathcal{O} \left( n \right) \)

Retreive the list of network edge identifiers present in the topology.

mutuallyExclusivePairs :: TopologyRepresentation a -> Set (a, a)

\( \mathcal{O} \left( n \right) \)

Retreive the list of network edge identifiers stored in the topology representation.

Comparison

isCompatableWithTopology :: Ord a => TopologyRepresentation a -> TopologyRepresentation a -> Bool

\( \mathcal{O} \left( m * \log_2 ( \frac {n}{m + 1} ) \right), m \leq n \)

Perform a subsetting operation to determine is a sub-topology is compatable with another topology.

isEdgePermissibleWith :: Ord a => a -> TopologyRepresentation a -> Bool

\( \mathcal{O} \left( \log_2 ( n ) \right) \)

Determine if an edge is compatible with a topology.