Text.XML

Description

Output tree structure and node decoration data to XML.

Documentation

class ToXML a where

(✔)

Minimal complete definition

toXML

Methods

toXML :: a -> Element

data Content :: *

XML content

Constructors

CRef String

collapseElemList :: (FoldableWithKey f, Show (Key f), ToXML a) => String -> [Attr] -> f a -> Element

Take in a list of ToXML items and return a single Element with XML'ed items as substructure. Used in ToXML instances when there are sequences of data.

ppTopElement :: Element -> String

Pretty printing renders XML documents faithfully, with the exception that whitespace may be added/removed in non-verbatim character data.

xmlAttr :: (String, String) -> Attr

Create an XML Attr, which is a key value pair (xmlQName, String).

xmlContent :: (String, String) -> Content

Create XML Contents, which for our use is always an Element.

xmlElement :: String -> [(String, String)] -> [Either (String, String) Element] -> Element

Create an XML Element from a String, a list of attributes and a list of contents. If the input is a tuple of strings, create an Element with the first as a tag and the second as content. Otherwise, return the Element val as Elem val (to create Content).

xmlQName :: String -> QName

Create a QName from a String.